% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{attach_existing_aml_compute}
\alias{attach_existing_aml_compute}
\title{Attach existing amlcompute}
\usage{
attach_existing_aml_compute(workspace, target_name)
}
\arguments{
\item{workspace}{workspace object}

\item{target_name}{name of the amlcompute}
}
\description{
Attach existing amlcompute
}
