% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{attach_existing_compute_target}
\alias{attach_existing_compute_target}
\title{Attach existing compute target}
\usage{
attach_existing_compute_target(workspace, target_name)
}
\arguments{
\item{workspace}{workspace object}

\item{target_name}{name of the amlcompute}
}
\description{
Attach existing compute target
}
