% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{attach_existing_dsvm}
\alias{attach_existing_dsvm}
\title{Attach existing dsvm}
\usage{
attach_existing_dsvm(workspace, target_name)
}
\arguments{
\item{workspace}{workspace object}

\item{target_name}{name of the dsvm}
}
\description{
Attach existing dsvm
}
