% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{bandit_policy}
\alias{bandit_policy}
\title{Create Bandit policy for HyperDrive runs}
\usage{
bandit_policy(slack_factor = NULL, slack_amount = NULL,
  evaluation_interval = 1, delay_evaluation = 0)
}
\arguments{
\item{slack_factor}{ratio of the allowed distance from best-performing run}

\item{slack_amount}{absolute allowed distance from the best-performing run}

\item{evaluation_interval}{frequency for applying policy}

\item{delay_evaluation}{how many intervals to delay the first evaluation}
}
\value{
EarlyTerminationPolicy object
}
\description{
Create Bandit policy for HyperDrive runs
}
