% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{bayesian_parameter_sampling}
\alias{bayesian_parameter_sampling}
\title{Define Bayesian Parameter sampling over hyperparameter search space}
\usage{
bayesian_parameter_sampling(parameter_space)
}
\arguments{
\item{parameter_space}{a named list containing each parameter and its
distribution}
}
\value{
HyperParameterSampling object
}
\description{
Define Bayesian Parameter sampling over hyperparameter search space
}
