% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{build_docker_image_name}
\alias{build_docker_image_name}
\title{Build the base docker image name incorporating the registry details.}
\usage{
build_docker_image_name(custom_docker_image = NULL,
  base_image_registry = NULL)
}
\arguments{
\item{custom_docker_image}{The name of the docker image from which the image to use for training will be built. If
not set, a default CPU based image will be used as the base image.}

\item{base_image_registry}{Image registry that contains the base image.}
}
\description{
Build the base docker image name incorporating the registry details.
}
