% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{cancel_run}
\alias{cancel_run}
\title{Cancel a run}
\usage{
cancel_run(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\value{
\code{TRUE} if cancellation was successful, else \code{FALSE}.
}
\description{
Cancel an ongoing run.
}
