% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{choice}
\alias{choice}
\title{Specify a discrete set of options to sample from}
\usage{
choice(options)
}
\arguments{
\item{options}{list of options to choose from}
}
\value{
the stochastic expression
}
\description{
Specify a discrete set of options to sample from
}
