% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{complete_run}
\alias{complete_run}
\title{Mark the run as completed. This is typically used in interactive logging
scenarios.}
\usage{
complete_run(run)
}
\arguments{
\item{run}{Run object.}
}
\description{
Mark the run as completed. This is typically used in interactive logging
scenarios.
}
