% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{container_registry}
\alias{container_registry}
\title{Get Azure Container Registry.}
\usage{
container_registry(address = NULL, username = NULL, password = NULL)
}
\arguments{
\item{address}{DNS name or IP address of azure container registry(ACR)}

\item{username}{The username for ACR}

\item{password}{The password for ACR}
}
\description{
Get Azure Container Registry.
}
