% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{convert_to_dataset_with_csv_files}
\alias{convert_to_dataset_with_csv_files}
\title{Convert the current dataset into a FileDataset containing CSV files.}
\usage{
convert_to_dataset_with_csv_files(dataset, separator = ",")
}
\arguments{
\item{dataset}{The Tabular Dataset object.}

\item{separator}{The separator to use to separate values in the resulting file.}
}
\value{
A new FileDataset object with a set of CSV files containing the data
in this dataset.
}
\description{
Convert the current dataset into a FileDataset containing CSV files.
}
