% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{convert_to_dataset_with_parquet_files}
\alias{convert_to_dataset_with_parquet_files}
\title{Convert the current dataset into a FileDataset containing Parquet files.}
\usage{
convert_to_dataset_with_parquet_files(dataset)
}
\arguments{
\item{dataset}{The Tabular Dataset object.}
}
\value{
A new FileDataset object with a set of Parquet files containing the
data in this dataset.
}
\description{
Convert the current dataset into a FileDataset containing Parquet files.
The resulting dataset will contain one or more Parquet files, each corresponding
to a partition of data from the current dataset. These files are not materialized
until they are downloaded or read from.
}
