% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{create_child_runs}
\alias{create_child_runs}
\title{Create one or many child runs}
\usage{
create_child_runs(parent_run, count = NULL, tag_key = NULL,
  tag_values = NULL)
}
\arguments{
\item{parent_run}{The \code{HyperDriveRun} object.}

\item{count}{An optional number of children to create.}

\item{tag_key}{An optional key to populate the Tags entry in all created children.}

\item{tag_values}{An optional list of values that will map onto Tags\link{tag_key}
for the list of runs created.}
}
\value{
The list of child runs.
}
\description{
Create one or many child runs.
}
