% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{create_conda_env}
\alias{create_conda_env}
\title{Create conda environment}
\usage{
create_conda_env(environment_name)
}
\arguments{
\item{environment_name}{name of environment to create}
}
\description{
Create conda environment
}
