% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reference.R
\name{create_data_reference}
\alias{create_data_reference}
\title{Makes a datastore available as a datareference for a run}
\usage{
create_data_reference(datastore, data_reference_name = NULL,
  path_on_datastore = NULL, mode = "mount", path_on_compute = NULL,
  overwrite = FALSE)
}
\arguments{
\item{datastore}{the Datastore to reference}

\item{data_reference_name}{the name of the data reference}

\item{path_on_datastore}{the relative path on cloud for the data reference}

\item{mode}{the operation on the data reference, we support mount, download}

\item{path_on_compute}{the path on compute for the data reference}

\item{overwrite}{overwrite if data reference exists}
}
\value{
data reference object
}
\description{
Makes a datastore available as a datareference for a run
}

