% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{create_dockerfile}
\alias{create_dockerfile}
\title{Creates a dockerfile which builds the image to run the entry_script.}
\usage{
create_dockerfile(custom_docker_image = NULL, cran_packages = NULL,
  github_packages = NULL, custom_url_packages = NULL)
}
\arguments{
\item{custom_docker_image}{The name of the docker image from which the image to use for training will be built. If
not set, a default CPU based image will be used as the base image.}

\item{cran_packages}{character vector of cran packages to be installed.}

\item{github_packages}{character vector of github packages to be installed.}

\item{custom_url_packages}{character vector of packages to be installed from local, directory or custom url.}
}
\description{
Creates a dockerfile which builds the image to run the entry_script.
}
