% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{create_file_dataset_from_files}
\alias{create_file_dataset_from_files}
\title{Create an unregistered, in-memory Dataset from files.}
\usage{
create_file_dataset_from_files(path, validate = TRUE)
}
\arguments{
\item{path}{A data path in a registered datastore or a local path.}

\item{validate}{Indicates whether to validate if data can be loaded from the
returned dataset. Defaults to True. Validation requires that the data source
is accessible from the current compute.}
}
\value{
The Dataset object
}
\description{
Create an unregistered, in-memory Dataset from files. Use this method to read
files as streams of data. Returns one file stream object per file read.
The returned Dataset is not registered with the workspace.
}
