% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{create_hyperdrive_config}
\alias{create_hyperdrive_config}
\title{Create configuration for a HyperDrive run}
\usage{
create_hyperdrive_config(estimator, param_sampling, policy,
  primary_metric_name, primary_metric_goal, max_total_runs,
  max_concurrent_runs = NULL)
}
\arguments{
\item{estimator}{estimator object}

\item{param_sampling}{parameter sampling option}

\item{policy}{termination policy}

\item{primary_metric_name}{name of primary metric}

\item{primary_metric_goal}{goal of primary metric}

\item{max_total_runs}{maximum number of runs to queue}

\item{max_concurrent_runs}{maximum number of runs to start concurrently}
}
\value{
HyperDrive config object
}
\description{
Create configuration for a HyperDrive run
}
