% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runconfig.R
\name{create_run_config}
\alias{create_run_config}
\title{Create run configuration to run a R script}
\usage{
create_run_config(target, data_references = NULL, base_image = NULL)
}
\arguments{
\item{target}{target string}

\item{data_references}{list of data references}

\item{base_image}{image to use}
}
\value{
runconfig object
}
\description{
Create run configuration to run a R script
}
