% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create_run_details_plot}
\alias{create_run_details_plot}
\title{Create run details plot}
\usage{
create_run_details_plot(run, rstudio_server)
}
\arguments{
\item{run}{run used for plotting}
}
\description{
Create run details plot
}
