% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data_path}
\alias{data_path}
\title{Represents a path to data in a datastore.}
\usage{
data_path(datastore = NULL, path_on_datastore = NULL, name = NULL)
}
\arguments{
\item{datastore}{The Datastore to reference.}

\item{path_on_datastore}{The relative path in the backing storage for the data reference.}

\item{name}{An optional name for the DataPath.}
}
\value{
The `DataPath`` object.
}
\description{
The path represented by DataPath object can point to a directory or a data artifact (blob, file).
}
\examples{
\dontrun{
path <- data_path(datastore, path_on_datastore, "my_path_1")

f_ds <- create_file_dataset_from_files(path)
}
}
\seealso{
\code{\link{create_file_dataset_from_files}}
\code{\link{create_tabular_dataset_from_parquet_files}}
\code{\link{create_tabular_dataset_from_delimited_files}}
\code{\link{create_tabular_dataset_from_json_lines_files}}
\code{\link{create_tabular_dataset_from_sql_query}}
}
