% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data_type_datetime}
\alias{data_type_datetime}
\title{Configure conversion to datetime.}
\usage{
data_type_datetime(formats = NULL)
}
\arguments{
\item{formats}{Formats to try for datetime conversion. For example \verb{\%d-\%m-\%Y} for data in "day-month-year",
and \verb{\%Y-\%m-\%dT\%H:\%M:\%S.\%f} for "combined date an time representation" according to ISO 8601.
\itemize{
\item \%Y: Year with 4 digits
\item \%y: Year with 2 digits
\item \%m: Month in digits
\item \%b: Month represented by its abbreviated name in 3 letters, like Aug
\item \%B: Month represented by its full name, like August
\item \%d: Day in digits
\item \%H: Hour as represented in 24-hour clock time
\item \%I: Hour as represented in 12-hour clock time
\item \%M: Minute in 2 digits
\item \%S: Second in 2 digits
\item \%f: Microsecond
\item \%p: AM/PM designator
\item \%z: Timezone, for example: -0700
}

Format specifiers will be inferred if not specified.
Inference requires that the data source is accessible from current compute.}
}
\value{
Converted DataType object.
}
\description{
Configure conversion to datetime.
}
