% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data_type_double}
\alias{data_type_double}
\title{Configure conversion to 53-bit double.}
\usage{
data_type_double()
}
\value{
Converted DataType object.
}
\description{
Configure conversion to 53-bit double.
}
