% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{data_type_long}
\alias{data_type_long}
\title{Configure conversion to 64-bit integer.}
\usage{
data_type_long()
}
\value{
Converted DataType object.
}
\description{
Configure conversion to 64-bit integer.
}
