% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{delete_compute}
\alias{delete_compute}
\title{Delete a cluster}
\usage{
delete_compute(cluster)
}
\arguments{
\item{cluster}{The \code{AmlCompute} or \code{AksCompute} object.}
}
\value{
None
}
\description{
Remove the compute object from its associated workspace and delete the
corresponding cloud-based resource.
}
\examples{
\dontrun{
ws <- load_workspace_from_config()
compute_target <- get_compute(ws, cluster_name = 'mycluster')
delete_compute(compute_target)
}
}
