% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-local.R
\name{delete_local_webservice}
\alias{delete_local_webservice}
\title{Delete this LocalWebservice from the local machine.
This function call is not asynchronous; it runs until the service is deleted.}
\usage{
delete_local_webservice(webservice, delete_cache = TRUE,
  delete_image = FALSE)
}
\arguments{
\item{webservice}{LocalWebservice object.}

\item{delete_cache}{Delete temporary files cached for the service.}

\item{delete_image}{Delete the service's Docker image.}
}
\description{
Delete this LocalWebservice from the local machine.
This function call is not asynchronous; it runs until the service is deleted.
}
