% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{delete_model}
\alias{delete_model}
\title{Delete this model from its associated workspace.}
\usage{
delete_model(model)
}
\arguments{
\item{model}{The model to download}
}
\description{
Delete this model from its associated workspace.
}
