% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{delete_secrets}
\alias{delete_secrets}
\title{Delete the secrets from the keyvault}
\usage{
delete_secrets(keyvault, secrets)
}
\arguments{
\item{keyvault}{keyvault object}

\item{secrets}{a vector of secret names}
}
\description{
Delete the secrets from the keyvault
}
