% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{delete_webservice}
\alias{delete_webservice}
\title{Delete this Webservice from its associated workspace.}
\usage{
delete_webservice(webservice)
}
\arguments{
\item{webservice}{The webservice object.}
}
\description{
Delete this Webservice from its associated workspace.
}
