% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{deploy_model}
\alias{deploy_model}
\title{Deploy a Webservice from zero or more model objects.}
\usage{
deploy_model(workspace, name, models, inference_config,
  deployment_config = NULL, deployment_target = NULL)
}
\arguments{
\item{workspace}{A Workspace object to associate the Webservice with}

\item{name}{The name to give the deployed service. Must be unique to the workspace, only consist of lowercase
letters, numbers, or dashes, start with a letter, and be between 3 and 32 characters long.}

\item{models}{A list of model objects. Can be an empty list.}

\item{inference_config}{An InferenceConfig object used to determine required model properties.}

\item{deployment_config}{A WebserviceDeploymentConfiguration used to configure the webservice. If one is not
provided, an empty configuration object will be used based on the desired target.}

\item{deployment_target}{A azureml.core.ComputeTarget to deploy the Webservice to.
As Azure Container Instances has no associated azureml.core.ComputeTarget, leave
this parameter as None to deploy to Azure Container Instances.}
}
\description{
Deploy a Webservice from zero or more model objects.
}
