% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{deserialize_to_model}
\alias{deserialize_to_model}
\title{Convert a json object into a Model object.}
\usage{
deserialize_to_model(workspace, model_payload)
}
\arguments{
\item{workspace}{The workspace object the model is registered under}

\item{model_payload}{A json object to convert to a Model object}
}
\description{
Convert a json object into a Model object.
}
