% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{deserialize_to_webservice}
\alias{deserialize_to_webservice}
\title{Convert a json object into a Webservice object.}
\usage{
deserialize_to_webservice(workspace, webservice_payload)
}
\arguments{
\item{workspace}{The workspace object the Webservice is registered under}

\item{webservice_payload}{A json object to convert to a Webservice object}
}
\description{
Convert a json object into a Webservice object.
}
