% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{detach_aks_compute}
\alias{detach_aks_compute}
\title{Detach the AksCompute object from its associated workspace}
\usage{
detach_aks_compute(cluster)
}
\arguments{
\item{cluster}{cluster object}
}
\description{
Detach the AksCompute object from its associated workspace
}
