% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{.create_run_details_plot}
\alias{.create_run_details_plot}
\title{Generate table of run details}
\usage{
.create_run_details_plot(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\description{
Generate table of run details
}
