% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{download_file_from_run}
\alias{download_file_from_run}
\title{Download an associated file from storage.}
\usage{
download_file_from_run(run, name, output_file_path = NULL)
}
\arguments{
\item{run}{the run object}

\item{name}{The name of the artifact to be downloaded}

\item{output_file_path}{The local path where to store the artifact}
}
\description{
Download an associated file from storage.
}
