% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{download_files_from_run}
\alias{download_files_from_run}
\title{Download files from a given storage prefix (folder name) or 
the entire container if prefix is unspecified.}
\usage{
download_files_from_run(run, prefix = NULL, output_directory = NULL,
  output_paths = NULL, batch_size = 100L)
}
\arguments{
\item{run}{the run object}

\item{prefix}{the filepath prefix within the container from 
which to download all artifacts}

\item{output_directory}{optional directory that all artifact paths use 
as a prefix}

\item{output_paths}{optional filepaths in which to store the downloaded 
artifacts. Should be unique and match length of paths.}

\item{batch_size}{number of files to download per batch}
}
\description{
Download files from a given storage prefix (folder name) or 
the entire container if prefix is unspecified.
}
