% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{download_from_datastore}
\alias{download_from_datastore}
\title{Download the data from the datastore to the local file system}
\usage{
download_from_datastore(ds, target_path, prefix = NULL, overwrite = FALSE,
  show_progress = TRUE)
}
\arguments{
\item{ds}{datastore object}

\item{target_path}{the local directory to download the file to}

\item{prefix}{path to the folder in the blob container to
download. If set to NULL, will download everything in the blob defaults to NULL}

\item{overwrite}{overwrite existing file, defaults to FALSE}

\item{show_progress}{show progress of download in the console, defaults to TRUE}
}
\description{
Download the data from the datastore to the local file system
}

