% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{download_from_file_dataset}
\alias{download_from_file_dataset}
\title{Download file streams defined by the dataset as local files.}
\usage{
download_from_file_dataset(dataset, target_path = NULL, overwrite = FALSE)
}
\arguments{
\item{dataset}{The Dataset object}

\item{target_path}{The local directory to download the files to. If NULL,
the data will be downloaded into a temporary directory.}

\item{overwrite}{Indicates whether to overwirte existing files. The default
is FALSE. Existing files will be overwritten if \code{overwrite} is set to TRUE;
otherwise an exception will be raised.}
}
\value{
A list of file paths for each file downloaded.
}
\description{
Download file streams defined by the dataset as local files. If target_path starts
with a /, then it will be treated as an absolute path. If it doesn't start with a /,
then it will be treated as a relative path relative to the current working directory.
}
