% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{download_model}
\alias{download_model}
\title{Download model to target_dir of local file system.}
\usage{
download_model(model, target_dir = ".", exist_ok = FALSE)
}
\arguments{
\item{model}{The model to download}

\item{target_dir}{Path to directory for where to download the model. Defaults to "."}

\item{exist_ok}{Boolean to replace downloaded dir/files if exists. Defaults to False}
}
\description{
Download model to target_dir of local file system.
}
