% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{drop_columns_from_dataset}
\alias{drop_columns_from_dataset}
\title{Drop the specified columns from the dataset.}
\usage{
drop_columns_from_dataset(dataset, columns)
}
\arguments{
\item{dataset}{The Tabular Dataset object.}

\item{columns}{A name or a list of names for the columns to drop.}
}
\value{
A new TabularDataset object with the specified columns dropped.
}
\description{
Drop the specified columns from the dataset. If a timeseries column is dropped,
the corresponding capabilities will be dropped for the returned dataset as well.
}
