% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{environment}
\alias{environment}
\title{Configure the python environment where the experiment is executed.}
\usage{
environment(name, version = NULL, environment_variables = NULL,
  cran_packages = NULL, github_packages = NULL,
  custom_url_packages = NULL, custom_docker_image = NULL,
  base_image_registry = NULL)
}
\arguments{
\item{name}{The name of the environment}

\item{version}{The version of the environment}

\item{environment_variables}{A dictionary of environment variables names and values.}

\item{cran_packages}{character vector of cran packages to be installed.}

\item{github_packages}{character vector of github packages to be installed.}

\item{custom_url_packages}{character vector of packages to be installed from local, directory or custom url.}

\item{custom_docker_image}{The name of the docker image from which the image to use for training will be built. If
not set, a default CPU based image will be used as the base image.}

\item{base_image_registry}{Image registry that contains the base image.}
}
\description{
Configure the python environment where the experiment is executed.
}
