% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{estimator}
\alias{estimator}
\title{Create an estimator}
\usage{
estimator(source_directory, compute_target = NULL, vm_size = NULL,
  vm_priority = NULL, entry_script = NULL, script_params = NULL,
  max_run_duration_seconds = NULL, environment = NULL, inputs = NULL)
}
\arguments{
\item{source_directory}{A string of the local directory containing
experiment configuration and code files needed for the training job.}

\item{compute_target}{The \code{AmlCompute} object for the compute target
where training will happen.}

\item{vm_size}{A string of the VM size of the compute target that will be
created for the training job. The list of available VM sizes
are listed \href{https://docs.microsoft.com/azure/cloud-services/cloud-services-sizes-specs}{here}.
Provide this parameter if you want to create AmlCompute as the compute target
at run time, instead of providing an existing cluster to the \code{compute_target}
parameter. If \code{vm_size} is specified, a single-node cluster is automatically
created for your run and is deleted automatically once the run completes.}

\item{vm_priority}{A string of either \code{'dedicated'} or \code{'lowpriority'} to
specify the VM priority of the compute target that will be created for the
training job. Defaults to \code{'dedicated'}. This takes effect only when the
\code{vm_size} parameter is specified.}

\item{entry_script}{A string representing the relative path to the file used
to start training.}

\item{script_params}{A named list of the command-line arguments to pass to
the training script specified in \code{entry_script}.}

\item{max_run_duration_seconds}{An integer of the maximum allowed time for
the run. Azure ML will attempt to automatically cancel the run if it takes
longer than this value.}

\item{environment}{The \code{Environment} object that configures the R
environment where the experiment is executed.}

\item{inputs}{A list of DataReference objects or DatasetConsumptionConfig
objects to use as input.}
}
\value{
The \code{Estimator} object.
}
\description{
An Estimator wraps run configuration information for specifying details
of executing an R script. Running an Estimator experiment
(using \code{submit_experiment()}) will return a \code{ScriptRun} object and
execute your training script on the specified compute target.

To define the environment to use for training, you can either directly
provide the environment-related parameters (e.g. \code{cran_packages},
\code{custom_docker_image}) to \code{estimator()}, or you can provide an
\code{Environment} object to the \code{environment} parameter. For more information
on the predefined Docker images that are used for training if
\code{custom_docker_image} is not specified, see the documentation
\href{https://azure.github.io/azureml-sdk-for-r/reference/r_environment.html#predefined-docker-images}{here}.
}
\section{Examples}{
\preformatted{est <- estimator(source_directory = ".",
                 entry_script = "train.R",
                 compute_target = compute_target)
}
}

\seealso{
\code{\link[=r_environment]{r_environment()}}, \code{\link[=container_registry]{container_registry()}}, \code{\link[=submit_experiment]{submit_experiment()}},
\code{\link[=dataset_consumption_config]{dataset_consumption_config()}}, \code{\link[=cran_package]{cran_package()}}
}
