% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Create an Azure Machine Learning Experiment}
\usage{
experiment(workspace, name)
}
\arguments{
\item{workspace}{The workspace object that would contain the experiment.}

\item{name}{The experiment name.}
}
\value{
experiment object
}
\description{
Create an Azure Machine Learning Experiment
}
