% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{filter_dataset_after_time}
\alias{filter_dataset_after_time}
\title{Filter Tabular Dataset with time stamp columns after a specified start time.}
\usage{
filter_dataset_after_time(dataset, start_time, include_boundary = TRUE)
}
\arguments{
\item{dataset}{The Tabular Dataset object}

\item{start_time}{The lower bound for filtering data.}

\item{include_boundary}{Boolean indicating if the row associated with the
boundary time (\code{start_time}) should be included.}
}
\value{
The filtered Tabular Dataset
}
\description{
Filter Tabular Dataset with time stamp columns after a specified start time.
}
