% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{filter_dataset_before_time}
\alias{filter_dataset_before_time}
\title{Filter Tabular Dataset with time stamp columns before a specified end time.}
\usage{
filter_dataset_before_time(dataset, end_time, include_boundary = TRUE)
}
\arguments{
\item{dataset}{The Tabular Dataset object}

\item{end_time}{The upper bound for filtering data.}

\item{include_boundary}{Boolean indicating if the row associated with the
boundary time (\code{start_time}) should be included.}
}
\value{
The filtered Tabular Dataset
}
\description{
Filter Tabular Dataset with time stamp columns before a specified end time.
}
