% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{filter_dataset_between_time}
\alias{filter_dataset_between_time}
\title{Filter Tabular Dataset between a specified start and end time.}
\usage{
filter_dataset_between_time(dataset, end_time, include_boundary = FALSE)
}
\arguments{
\item{dataset}{The Tabular Dataset object}

\item{end_time}{The upper bound for filtering data.}

\item{include_boundary}{Boolean indicating if the row associated with the
boundary time (\code{start_time} and \code{end_time}) should be included.}

\item{start_time}{The lower bound for filtering data.}
}
\value{
The filtered Tabular Dataset
}
\description{
Filter Tabular Dataset between a specified start and end time.
}
