% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{filter_dataset_from_recent_time}
\alias{filter_dataset_from_recent_time}
\title{Filter Tabular Dataset to contain only the specified duration (amount) of recent data.}
\usage{
filter_dataset_from_recent_time(dataset, time_delta, include_boundary = TRUE)
}
\arguments{
\item{dataset}{The Tabular Dataset object}

\item{time_delta}{The duration (amount) of recent data to retrieve.}

\item{include_boundary}{Boolean indicating if the row associated with the
boundary time (\code{time_delta}) should be included.}
}
\value{
The filtered Tabular Dataset
}
\description{
Filter Tabular Dataset to contain only the specified duration (amount) of recent data.
}
