% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_azureml_parallel.R
\name{generate_entry_script}
\alias{generate_entry_script}
\title{Generates the control script for the experiment.}
\usage{
generate_entry_script(source_directory)
}
\arguments{
\item{source_directory}{The directory which contains all the files
needed for the experiment.}
}
\description{
Generates the control script for the experiment.
}
