% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{generate_new_webservice_key}
\alias{generate_new_webservice_key}
\title{Regenerate one of the Webservice's keys. Must specify either 'Primary' or
'Secondary' key.}
\usage{
generate_new_webservice_key(webservice, key_type)
}
\arguments{
\item{webservice}{The webservice object.}

\item{key_type}{Which key to regenerate. Options are 'Primary' or 
'Secondary'}
}
\description{
Regenerate one of the Webservice's keys. Must specify either 'Primary' or 
'Secondary' key.
}
