% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{generate_score_python_wrapper}
\alias{generate_score_python_wrapper}
\title{Generate _generated_score.py file for the corresponding entry_script file}
\usage{
generate_score_python_wrapper(entry_script, source_directory)
}
\arguments{
\item{entry_script}{Path to local file that contains the code to run for
the image.}

\item{source_directory}{paths to folders that contains all files to
create the image.}
}
\description{
Generate _generated_score.py file for the corresponding entry_script file
}
