% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_aks_compute_credentials}
\alias{get_aks_compute_credentials}
\title{Get the credentials for an AksCompute cluster}
\usage{
get_aks_compute_credentials(cluster)
}
\arguments{
\item{cluster}{The \code{AksCompute} object.}
}
\value{
Details of the cluster credentials as data frame.
}
\description{
Retrieved the credentials for an AksCompute cluster.
}
