% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{get_best_run_by_primary_metric}
\alias{get_best_run_by_primary_metric}
\title{Return the best performing run amongst all completed runs}
\usage{
get_best_run_by_primary_metric(hyperdrive_run, include_failed = TRUE,
  include_canceled = TRUE)
}
\arguments{
\item{hyperdrive_run}{HyperDriveRun object}

\item{include_failed}{whether to include failed runs}

\item{include_canceled}{whether to include canceled runs}
}
\value{
Run object
}
\description{
Return the best performing run amongst all completed runs
}
