% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{get_child_run_hyperparameters}
\alias{get_child_run_hyperparameters}
\title{Return hyperparameters for all child runs}
\usage{
get_child_run_hyperparameters(hyperdrive_run)
}
\arguments{
\item{hyperdrive_run}{HyperDriveRun object}
}
\value{
named list of hyperparameters grouped by run_id
}
\description{
Return hyperparameters for all child runs
}
