% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_run.R
\name{get_children_hyperparameters}
\alias{get_children_hyperparameters}
\title{Return hyperparameters for all child runs}
\usage{
get_children_hyperparameters(hyperdrive_run)
}
\arguments{
\item{hyperdrive_run}{HyperDriveRun object}
}
\value{
named list of hyperparameters grouped by run_id
}
\description{
Return hyperparameters for all child runs
}
