% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{get_children_metrics}
\alias{get_children_metrics}
\title{Return metrics from all child runs}
\usage{
get_children_metrics(hyperdrive_run)
}
\arguments{
\item{hyperdrive_run}{HyperDriveRun object}
}
\value{
name list of metrics grouped by run_id
}
\description{
Return metrics from all child runs
}
