% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_run.R
\name{get_children_sorted_by_primary_metric}
\alias{get_children_sorted_by_primary_metric}
\title{Return the child runs sorted in descending order by best primary metric}
\usage{
get_children_sorted_by_primary_metric(hyperdrive_run, top = 0,
  reverse = FALSE, discard_no_metric = FALSE)
}
\arguments{
\item{hyperdrive_run}{HyperDriveRun object}

\item{top}{number of top children to be returned, deafult value of 0 will return all}

\item{reverse}{reverse the sorting order}

\item{discard_no_metric}{whether to include children without the primary metric}
}
\value{
named list of child runs
}
\description{
Return the child runs sorted in descending order by best primary metric
}
