% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_compute}
\alias{get_compute}
\title{Get amlcompute. Returns NULL if amlcompute is not found on workspace.}
\usage{
get_compute(workspace, cluster_name)
}
\arguments{
\item{workspace}{workspace that has the cluster}

\item{cluster_name}{name of the cluster}
}
\description{
Get amlcompute. Returns NULL if amlcompute is not found on workspace.
}

