% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_current_run}
\alias{get_current_run}
\title{Gets the context object for a run}
\usage{
get_current_run(allow_offline = TRUE)
}
\arguments{
\item{allow_offline}{Allow the service context to fall back to offline mode 
so that the training script
can be tested locally without submitting a job with the SDK.}
}
\value{
get current run
}
\description{
Gets the context object for a run
}
