% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reference.R
\name{get_data_reference_configuration}
\alias{get_data_reference_configuration}
\title{Create data reference configuration}
\usage{
get_data_reference_configuration(data_reference)
}
\arguments{
\item{data_reference}{Data reference object}
}
\description{
Create data reference configuration
}
