% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_reference.R
\name{get_data_reference_path_in_compute}
\alias{get_data_reference_path_in_compute}
\title{Get data reference path on path_on_compute}
\usage{
get_data_reference_path_in_compute(data_reference)
}
\arguments{
\item{data_reference}{DataReference whose path is needed}
}
\value{
environment variable in remote compute that holds the data reference path.
}
\description{
Get data reference path on path_on_compute
}

