% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_dataset_by_name}
\alias{get_dataset_by_name}
\title{Get a registered Dataset from the workspace by its registration name.}
\usage{
get_dataset_by_name(workspace, name, version = "latest")
}
\arguments{
\item{workspace}{The existing AzureML workspace in which the Dataset was registered.}

\item{name}{The registration name.}

\item{version}{The registration version. Defaults to "latest".}
}
\value{
The registered Dataset object.
}
\description{
Get a registered Dataset from the workspace by its registration name.
}
