% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{get_datastore}
\alias{get_datastore}
\title{Get an existing datastore}
\usage{
get_datastore(workspace, datastore_name)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{datastore_name}{A string of the name of the datastore.}
}
\value{
The \code{AzureBlobDatastore} or \code{AzureFileDatastore} object.
}
\description{
Get the corresponding datastore object for an existing
datastore by name from the given workspace.
}
