% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_default_datastore}
\alias{get_default_datastore}
\title{Get default datastore associated with a workspace}
\usage{
get_default_datastore(workspace)
}
\arguments{
\item{workspace}{The workspace object}
}
\value{
The default datastore
}
\description{
Get default datastore associated with a workspace
}
