% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_default_keyvault}
\alias{get_default_keyvault}
\title{Get the default keyvault object for the workspace.}
\usage{
get_default_keyvault(workspace)
}
\arguments{
\item{workspace}{The workspace object}
}
\value{
The keyvault object associated with your workspace
}
\description{
Get the default keyvault object for the workspace.
}
