% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_environment}
\alias{get_environment}
\title{Return the environment object.}
\usage{
get_environment(workspace, name, version = NULL)
}
\arguments{
\item{workspace}{The workspace}

\item{name}{The name of the environment}

\item{version}{The version of the environment}
}
\description{
Return the environment object.
}
