% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_file_dataset_paths}
\alias{get_file_dataset_paths}
\title{Get a list of file paths for each file stream defined by the dataset.}
\usage{
get_file_dataset_paths(dataset)
}
\arguments{
\item{dataset}{The Dataset object.}
}
\value{
A list of file paths.
}
\description{
Get a list of file paths for each file stream defined by the dataset. The file
paths are relative paths for local files when the file srteam are downloaded
or mounted. A common prefix will be removed from the file paths based on how
data source was specified to create the dataset.
}
