% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_input_dataset_from_run}
\alias{get_input_dataset_from_run}
\title{Return the named list for input datasets.}
\usage{
get_input_dataset_from_run(name, run)
}
\arguments{
\item{name}{The name of the input dataset}

\item{run}{The run taking the dataset as input}
}
\value{
A dictionary with the name being the dataset's name and value being
the delivered data. If the mode is set to mount or download, it will return
the base path of the delivered data. If the mode is set to direct, it will
return the dataset object.
}
\description{
Return the named list for input datasets.
}
