% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_package_container_registry}
\alias{get_model_package_container_registry}
\title{Return a ContainerRegistry object for where the image
(or base image, for Dockerfile packages) is stored.}
\usage{
get_model_package_container_registry(package)
}
\arguments{
\item{package}{Package created with model(s) and dependencies.}
}
\value{
ContainerRegistry object
}
\description{
Return a ContainerRegistry object for where the image
(or base image, for Dockerfile packages) is stored.
}
