% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_package_creation_logs}
\alias{get_model_package_creation_logs}
\title{Retrieve the package creation logs.}
\usage{
get_model_package_creation_logs(package, decode = TRUE, offset = 0)
}
\arguments{
\item{package}{Package created with model(s) and dependencies.}

\item{decode}{Whether to decode the raw log bytes to a string.}

\item{offset}{Byte offset from which to start reading the logs.}
}
\value{
Package creation logs.
}
\description{
Retrieve the package creation logs.
}
