% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{get_param_sampling}
\alias{get_param_sampling}
\title{Get early termination policy for HyperDrive runs}
\usage{
get_param_sampling(sampling_type)
}
\arguments{
\item{policy_type}{one of "random", "grid", "bayesian"}
}
\value{
Parameter sampling object
}
\description{
Get early termination policy for HyperDrive runs
}
