% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run}
\alias{get_run}
\title{Gets the Run object from a given run id}
\usage{
get_run(experiment, run_id)
}
\arguments{
\item{experiment}{The containing experiment.}

\item{run_id}{The run id for the run.}
}
\value{
The run object.
}
\description{
Gets the Run object from a given run id
}
