% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run_details}
\alias{get_run_details}
\title{Get the definition, status information, current log files and other details
of the run.}
\usage{
get_run_details(run)
}
\arguments{
\item{run}{the run object}
}
\value{
Return the details for the run
}
\description{
Get the definition, status information, current log files and other details
of the run.
}
