% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run_details_with_logs}
\alias{get_run_details_with_logs}
\title{Return run status including log file content.}
\usage{
get_run_details_with_logs(run)
}
\arguments{
\item{run}{the run object}
}
\value{
Returns the status for the run with log file contents
}
\description{
Return run status including log file content.
}
