% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run_file_names}
\alias{get_run_file_names}
\title{List the files that are stored in association with the run.}
\usage{
get_run_file_names(run)
}
\arguments{
\item{run}{the run object}
}
\value{
The list of paths for existing artifacts
}
\description{
List the files that are stored in association with the run.
}
