% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{get_secrets}
\alias{get_secrets}
\title{Return the secret values for a given vector of secret names}
\usage{
get_secrets(keyvault, secrets)
}
\arguments{
\item{keyvault}{keyvault object}

\item{secrets}{a vector of secret names}
}
\value{
Returns a list of found and not found secrets
}
\description{
Return the secret values for a given vector of secret names
}
