% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_secrets_from_run}
\alias{get_secrets_from_run}
\title{Get the secret values for a given list of secret names.
Get a dictionary of found and not found secrets for the list of names 
provided.}
\usage{
get_secrets_from_run(run, secrets)
}
\arguments{
\item{run}{the run object}

\item{secrets}{List of secret names to retrieve the values for}
}
\value{
Returns a dictionary of found and not found secrets
}
\description{
Get the secret values for a given list of secret names.
Get a dictionary of found and not found secrets for the list of names 
provided.
}
