% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{get_termination_policy}
\alias{get_termination_policy}
\title{Get early termination policy for HyperDrive runs}
\usage{
get_termination_policy(policy_type)
}
\arguments{
\item{policy_type}{one of "truncation", "bandit", "median_stopping", "no_policy"}
}
\value{
Early Termination Policy object
}
\description{
Get early termination policy for HyperDrive runs
}
