% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{get_webservice_keys}
\alias{get_webservice_keys}
\title{Retrieve auth keys for this Webservice.}
\usage{
get_webservice_keys(webservice)
}
\arguments{
\item{webservice}{The webservice object.}
}
\description{
Retrieve auth keys for this Webservice.
}
