% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{get_webservice_logs}
\alias{get_webservice_logs}
\title{Retrieve logs for the Webservice.}
\usage{
get_webservice_logs(webservice, num_lines = 5000)
}
\arguments{
\item{webservice}{The webservice object.}

\item{num_lines}{The maximum number of log lines to retrieve.}
}
\description{
Retrieve logs for the Webservice.
}
