% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{get_webservice_token}
\alias{get_webservice_token}
\title{Retrieve auth token for this Webservice, scoped to the current user.}
\usage{
get_webservice_token(webservice)
}
\arguments{
\item{webservice}{The webservice object.}
}
\description{
Retrieve auth token for this Webservice, scoped to the current user.
}
