% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_workspace_details}
\alias{get_workspace_details}
\title{Return details of this workspace.}
\usage{
get_workspace_details(workspace)
}
\arguments{
\item{workspace}{The workspace object}
}
\value{
Workspace details
}
\description{
Return details of this workspace.
}
