% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{github_package}
\alias{github_package}
\title{Specifies a Github package to install in environment}
\usage{
github_package(repository, auth_token = NULL)
}
\arguments{
\item{repository}{Repository address of the github package}

\item{auth_token}{Personal access token to install from a private repo.}
}
\value{
A named list containing the package specifications
}
\description{
Specifies a Github package to install in run environment
}
\section{Examples}{
\preformatted{pkg1 <- github_package("Azure/azureml-sdk-for-r")

env <- r_environment(name = "r_env",
                     github_packages = list(pkg1))
}
}

\seealso{
\code{\link[=r_environment]{r_environment()}}
}
