% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{grid_parameter_sampling}
\alias{grid_parameter_sampling}
\title{Define Grid Parameter sampling over hyperparameter search space}
\usage{
grid_parameter_sampling(parameter_space)
}
\arguments{
\item{parameter_space}{a named list containing each parameter and its distribution}
}
\value{
HyperParameterSampling object
}
\description{
Define Grid Parameter sampling over hyperparameter search space
}
