% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{hyperdrive_config}
\alias{hyperdrive_config}
\title{Create configuration for a HyperDrive run}
\usage{
hyperdrive_config(hyperparameter_sampling, primary_metric_name,
  primary_metric_goal, max_total_runs, max_concurrent_runs = NULL,
  max_duration_minutes = 10080L, policy = NULL, estimator = NULL)
}
\arguments{
\item{hyperparameter_sampling}{hyperparameter sampling option}

\item{primary_metric_name}{name of primary metric}

\item{primary_metric_goal}{goal of primary metric}

\item{max_total_runs}{maximum number of runs to queue}

\item{max_concurrent_runs}{maximum number of runs to start concurrently}

\item{max_duration_minutes}{maximum time to allow runs}

\item{policy}{termination policy}

\item{estimator}{Estimator object}
}
\value{
HyperDriveConfig object
}
\description{
Create configuration for a HyperDrive run
}
