% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_azureml}
\alias{install_azureml}
\title{Install azureml}
\usage{
install_azureml(version = NULL, environment_name = "r-azureml")
}
\arguments{
\item{version}{pip version}

\item{environment_name}{name of environment to create}
}
\description{
Install azureml
}
