% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{keep_columns_from_dataset}
\alias{keep_columns_from_dataset}
\title{Keep the specified columns and drops all others from the dataset.}
\usage{
keep_columns_from_dataset(dataset, columns, validate = FALSE)
}
\arguments{
\item{dataset}{The Tabular Dataset object}

\item{columns}{The name or a list of names for the columns to keep.}

\item{validate}{Indicates whether to validate if data can be loaded from the
returned dataset. The default is False. Validation requires that the data
source is accessible from current compute.}
}
\value{
A new Tabular Dataset object with only the specified columns kept.
}
\description{
Keep the specified columns and drops all others from the dataset.
If a timeseries column is dropped, the corresponding capabilities will be
dropped for the returned dataset as well.
}
