% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{list_environments_in_workspace}
\alias{list_environments_in_workspace}
\title{Return the list of environments in the workspace.}
\usage{
list_environments_in_workspace(workspace)
}
\arguments{
\item{workspace}{The workspace}
}
\description{
Return the list of environments in the workspace.
}
