% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{list_nodes_in_aml_compute}
\alias{list_nodes_in_aml_compute}
\title{Get the details (e.g IP address, port etc) of all the compute nodes in the
compute target}
\usage{
list_nodes_in_aml_compute(cluster)
}
\arguments{
\item{cluster}{cluster object}
}
\value{
Details of all the compute nodes in the cluster in data frame
}
\description{
Get the details (e.g IP address, port etc) of all the compute nodes in the
compute target
}
