% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{list_secrets}
\alias{list_secrets}
\title{Return the list of secret names}
\usage{
list_secrets(keyvault)
}
\arguments{
\item{keyvault}{keyvault object}
}
\value{
Returns secret names for a given keyvault
}
\description{
Return the list of secret names
}
