% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{load_dataset_into_data_frame}
\alias{load_dataset_into_data_frame}
\title{Load all records from the dataset into a pandas DataFrame.}
\usage{
load_dataset_into_data_frame(dataset)
}
\arguments{
\item{dataset}{The Tabular Dataset object.}
}
\value{
A pandas DataFrame.
}
\description{
Load all records from the dataset into a pandas DataFrame.
}
