% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-local.R
\name{local_webservice_deployment_config}
\alias{local_webservice_deployment_config}
\title{Create a configuration object for deploying a local Webservice.}
\usage{
local_webservice_deployment_config(port = NULL)
}
\arguments{
\item{port}{The local port on which to expose the service's HTTP endpoint.}
}
\value{
LocalWebserviceDeploymentConfiguration object to use when deploying 
a Webservice object.
}
\description{
Create a configuration object for deploying a local Webservice.
}
