% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_confusion_matrix_to_run}
\alias{log_confusion_matrix_to_run}
\title{Log a confusion matrix to the artifact store.}
\usage{
log_confusion_matrix_to_run(name, value, description = "", run = NULL)
}
\arguments{
\item{name}{The name of the confusion matrix}

\item{value}{json containing name, version, and data properties}

\item{description}{An optional metric description}

\item{run}{Run object. If not specified, will default to current run from 
service context.}
}
\description{
Log a confusion matrix to the artifact store.
}
