% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_image_to_run}
\alias{log_image_to_run}
\title{Log an image metric to the run record.}
\usage{
log_image_to_run(name, path = NULL, plot = NULL, description = "",
  run = NULL)
}
\arguments{
\item{name}{The name of metric}

\item{path}{The path or stream of the image}

\item{plot}{The ggplot to log as an image.}

\item{description}{An optional metric description}

\item{run}{Run object. If not specified, will default to current run from
service context.}
}
\description{
Log an image metric to the run record.
}
