% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_list_to_run}
\alias{log_list_to_run}
\title{Log a list metric value to the run with the given name.}
\usage{
log_list_to_run(name, value, description = "", run = NULL)
}
\arguments{
\item{name}{The name of metric}

\item{value}{The value of the metric}

\item{description}{An optional metric description}

\item{run}{Run object. If not specified, will default to current run from 
service context.}
}
\description{
Log a list metric value to the run with the given name.
}
