% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_metric_to_run}
\alias{log_metric_to_run}
\title{Log metric to run}
\usage{
log_metric_to_run(name, value, run)
}
\arguments{
\item{name}{name of the metric}

\item{value}{value of the metric}

\item{run}{Run object}
}
\description{
Log metric to run
}
