% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_row_to_run}
\alias{log_row_to_run}
\title{Log a row metric to the run with the given name.}
\usage{
log_row_to_run(name, description = "", run = NULL, ...)
}
\arguments{
\item{name}{The name of metric}

\item{description}{An optional metric description}

\item{run}{Run object. If not specified, will default to current run from
service context.}

\item{...}{Each named parameter generates a column with the value specified.}
}
\description{
Log a row metric to the run with the given name.
}
