% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_table_to_run}
\alias{log_table_to_run}
\title{Log a table metric to the run with the given name.}
\usage{
log_table_to_run(name, value, description = "", run = NULL)
}
\arguments{
\item{name}{The name of metric}

\item{value}{The table value of the metric (dictionary where keys are
columns to be posted to the service)}

\item{description}{An optional metric description}

\item{run}{Run object. If not specified, will default to current run from
service context.}
}
\description{
Log a table metric to the run with the given name.
}
