% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{lognormal}
\alias{lognormal}
\title{Specify a normal distribution of the form exp(normal(mu, sigma))}
\usage{
lognormal(mu, sigma)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}
}
\value{
the stochastic expression
}
\description{
Specify a normal distribution of the form exp(normal(mu, sigma))
}
