% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{loguniform}
\alias{loguniform}
\title{Specify a log uniform distribution}
\usage{
loguniform(min_value, max_value)
}
\arguments{
\item{min_value}{minimum value in the range will be exp(min_value) (inclusive)}

\item{max_value}{maximum value in the range will be exp(min_value) (inclusive)}
}
\value{
the stochastic expression
}
\description{
Specify a log uniform distribution
}
