% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_azureml_parallel.R
\name{merge_results}
\alias{merge_results}
\title{Combine the results from the parallel training.}
\usage{
merge_results(node_count, process_count_per_node, run, source_directory)
}
\arguments{
\item{node_count}{Number of nodes in the AmlCompute cluster.}

\item{process_count_per_node}{Number of processes per node.}

\item{run}{The run object whose output needs to be combined.}

\item{source_directory}{The directory where the output from the run
would be downloaded.}
}
\description{
Combine the results from the parallel training.
}
