% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{normal}
\alias{normal}
\title{Specify a real value that is normally-distributed with mean mu and standard
deviation sigma}
\usage{
normal(mu, sigma)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}
}
\value{
the stochastic expression
}
\description{
Specify a real value that is normally-distributed with mean mu and standard
deviation sigma
}
