% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{primary_metric_goal}
\alias{primary_metric_goal}
\title{Define supported metric goals for hyperparameter tuning}
\usage{
primary_metric_goal(goal)
}
\arguments{
\item{goal}{string name of the goal ("MAXIMIZE" or "MINIMIZE")}
}
\value{
PrimaryMetricGoal object
}
\description{
Define supported metric goals for hyperparameter tuning
}
