% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{promote_headers_behavior}
\alias{promote_headers_behavior}
\title{Defines options for how column headers are processed when reading data from files to create a dataset.}
\usage{
promote_headers_behavior(behavior)
}
\arguments{
\item{option}{A string describing how column headers are to be processed
\itemize{
\item NO_HEADERS No column headers are read
\item ONLY_FIRST_FILE_HAS_HEADERS Read headers only from first row of first file, everything else is data.
\item COMBINE_ALL_FILES_HEADERS Read headers from first row of each file, combining identically named columns.
\item ALL_FILES_HAVE_SAME_HEADERS Read headers from first row of first file, drops first row from other files.
}}
}
\value{
The PromoteHeadersBehavior object.
}
\description{
Defines options for how column headers are processed when reading data from files to create a dataset.
These enumeration values are used in the Dataset class method.
}
