% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pull_model_package_image}
\alias{pull_model_package_image}
\title{Pull the package output to the local machine.
This can only be used with a Docker image package.}
\usage{
pull_model_package_image(package)
}
\arguments{
\item{package}{Package created with model(s) and dependencies.}
}
\description{
Pull the package output to the local machine.
This can only be used with a Docker image package.
}
