% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{qlognormal}
\alias{qlognormal}
\title{Specify a normal distribution of the form round(exp(normal(mu, sigma)) / q) * q}
\usage{
qlognormal(mu, sigma, q)
}
\arguments{
\item{mu}{mean of the normal distribution}

\item{sigma}{standard deviation of the normal distribution}

\item{q}{smoothing factor}
}
\value{
the stochastic expression
}
\description{
Specify a normal distribution of the form round(exp(normal(mu, sigma)) / q) * q
}
