% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{quniform}
\alias{quniform}
\title{Specify a uniform distribution of the form round(uniform(min_value, max_value) / q) * q}
\usage{
quniform(min_value, max_value, q)
}
\arguments{
\item{min_value}{minimum value in the range (inclusive)}

\item{max_value}{maximum value in the range (inclusive)}

\item{q}{smoothing factor}
}
\value{
the stochastic expression
}
\description{
Specify a uniform distribution of the form round(uniform(min_value, max_value) / q) * q
}
