% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{r_environment}
\alias{r_environment}
\title{Configure the R environment where the experiment is executed.}
\usage{
r_environment(name, version = NULL, environment_variables = NULL,
  cran_packages = NULL, github_packages = NULL,
  custom_url_packages = NULL, custom_docker_image = NULL,
  image_registry_details = NULL, use_gpu = FALSE, shm_size = NULL)
}
\arguments{
\item{name}{The name of the environment}

\item{version}{The version of the environment}

\item{environment_variables}{A dictionary of environment variables names and
values.}

\item{cran_packages}{character vector of cran packages to be installed.}

\item{github_packages}{character vector of github packages to be installed.}

\item{custom_url_packages}{character vector of packages to be installed from
local, directory or custom url.}

\item{custom_docker_image}{The name of the docker image from which the image
to use for training will be built. If not set, a default CPU based image will
be used as the base image.}

\item{image_registry_details}{The details of the Docker image registry.}

\item{use_gpu}{Indicates whether the environment to run the experiment should
support GPUs.
If TRUE, a GPU - based default Docker image will be used in the environment.
If FALSE, a CPU - based image will be used. Default Docker images
(CPU or GPU) will be used only if the 'custom_docker_image' parameter is not
set.}

\item{shm_size}{The size of the Docker container's shared memory block.}
}
\value{
the environment object.
}
\description{
Configure the R environment where the experiment is executed.
}