% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{randint}
\alias{randint}
\title{Specify a set of random integers in the range [0, upper)}
\usage{
randint(upper)
}
\arguments{
\item{upper}{upper bound for the range of integers (exclusive)}
}
\value{
the stochastic expression
}
\description{
Specify a set of random integers in the range [0, upper)
}
