% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{random_parameter_sampling}
\alias{random_parameter_sampling}
\title{Define Random Parameter sampling over hyperparameter search space}
\usage{
random_parameter_sampling(parameter_space, properties = NULL)
}
\arguments{
\item{parameter_space}{a named list containing each parameter and its distribution}

\item{properties}{a named list of additional properties for the algorithm}
}
\value{
HyperParameterSampling object
}
\description{
Define Random Parameter sampling over hyperparameter search space
}
