% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{random_split_dataset}
\alias{random_split_dataset}
\title{Split file streams in the dataset into two parts randomly and approximately by the percentage specified.}
\usage{
random_split_dataset(dataset, percentage, seed = NULL)
}
\arguments{
\item{dataset}{The Dataset object.}

\item{percentage}{The approximate percentage to split the Dataset by. This must
be a number between 0.0 and 1.0.}

\item{seed}{An optional seed to use for the random generator.}
}
\value{
A new Dataset object representing the two datasets after the split.
}
\description{
Split file streams in the dataset into two parts randomly and approximately by the percentage specified.
}
