% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{register_azure_blob_container_datastore}
\alias{register_azure_blob_container_datastore}
\title{Register an Azure Blob Container to the datastore.
You can choose to use SAS Token or Storage Account Key}
\usage{
register_azure_blob_container_datastore(workspace, datastore_name,
  container_name, account_name, sas_token = NULL, account_key = NULL,
  protocol = NULL, endpoint = NULL, overwrite = FALSE,
  create_if_not_exists = FALSE, skip_validation = FALSE,
  blob_cache_timeout = NULL, grant_workspace_access = FALSE,
  subscription_id = NULL, resource_group = NULL)
}
\arguments{
\item{workspace}{The workspace object}

\item{datastore_name}{The name of the datastore, case insensitive, can only 
contain alphanumeric characters and _}

\item{container_name}{The name of the azure blob container.}

\item{account_name}{The storage account name.}

\item{sas_token}{An account SAS token, defaults to NULL.}

\item{account_key}{A storage account key, defaults to NULL.}

\item{protocol}{Protocol to use to connect to the blob container. If NULL, 
defaults to https.}

\item{endpoint}{The endpoint of the blob container. If NULL, defaults to 
core.windows.net.}

\item{overwrite}{overwrites an existing datastore. If the datastore does not 
exist, it will create one, defaults to FALSE}

\item{create_if_not_exists}{create the file share if it does not exists, 
defaults to FALSE}

\item{skip_validation}{skips validation of storage keys, defaults to FALSE}

\item{blob_cache_timeout}{When this blob is mounted, set the cache timeout 
to this many seconds. If NULL, defaults to no timeout (i.e. blobs will be 
cached for the duration of the job when read).}

\item{grant_workspace_access}{grants Workspace Managed Identities(MSI) access 
to the user storage account, defaults to FALSE This should be set if the 
Storage account is in VNET. If set to TRUE, we will use the Workspace MSI 
token to grant access to the user storage account. It may take a while for 
the granted access to reflect.}

\item{subscription_id}{The subscription id of the storage account, defaults 
to NULL.}

\item{resource_group}{The resource group of the storage account, defaults 
to NULL.}
}
\value{
The blob datastore.
}
\description{
Register an Azure Blob Container to the datastore.
You can choose to use SAS Token or Storage Account Key
}
