% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{register_azure_file_share_datastore}
\alias{register_azure_file_share_datastore}
\title{Register an Azure File Share to the datastore.
You can choose to use SAS Token or Storage Account Key}
\usage{
register_azure_file_share_datastore(workspace, datastore_name,
  file_share_name, account_name, sas_token = NULL, account_key = NULL,
  protocol = NULL, endpoint = NULL, overwrite = FALSE,
  create_if_not_exists = FALSE, skip_validation = FALSE)
}
\arguments{
\item{workspace}{The workspace object}

\item{datastore_name}{The name of the datastore, case insensitive, can only 
contain alphanumeric characters and _}

\item{file_share_name}{The name of the azure file container.}

\item{account_name}{The storage account name.}

\item{sas_token}{An account SAS token, defaults to NULL.}

\item{account_key}{A storage account key, defaults to NULL.}

\item{protocol}{Protocol to use to connect to the blob container. 
If NULL, defaults to https.}

\item{endpoint}{The endpoint of the blob container. 
If NULL, defaults to core.windows.net.}

\item{overwrite}{overwrites an existing datastore. If the datastore does not 
exist, it will create one, defaults to FALSE}

\item{create_if_not_exists}{create the file share if it does not exists, 
defaults to FALSE}

\item{skip_validation}{skips validation of storage keys, defaults to FALSE}
}
\value{
The file datastore.
}
\description{
Register an Azure File Share to the datastore.
You can choose to use SAS Token or Storage Account Key
}
