% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{register_dataset}
\alias{register_dataset}
\title{Register a Dataset in the workspace}
\usage{
register_dataset(
  workspace,
  dataset,
  name,
  description = NULL,
  tags = NULL,
  create_new_version = FALSE
)
}
\arguments{
\item{workspace}{The AzureML workspace in which the Dataset is to be registered.}

\item{dataset}{The dataset to be registered.}

\item{name}{The name of the Dataset in the workspace.}

\item{description}{A description of the Dataset.}

\item{tags}{Named list of tags to give the Dataset. Defaults to NULL.}

\item{create_new_version}{Boolean to register the dataset as a new version under the specified name.}
}
\value{
The registered Dataset object.
}
\description{
Register the Dataset in the workspace, making it available to other users of the workspace.
}
