% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_azureml_parallel.R
\name{register_do_azureml_parallel}
\alias{register_do_azureml_parallel}
\title{Registers AMLCompute as a parallel backend with the foreach package.}
\usage{
register_do_azureml_parallel(workspace, compute_target)
}
\arguments{
\item{workspace}{The Workspace object which has the compute_target.}

\item{compute_target}{The AMLCompute target to use for parallelization.}
}
\description{
Registers AMLCompute as a parallel backend with the foreach package.
}
