% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{register_environment}
\alias{register_environment}
\title{Returns the environment object}
\usage{
register_environment(environment, workspace)
}
\arguments{
\item{environment}{The python environment where the experiment is executed.}

\item{workspace}{The workspace}
}
\description{
Returns the environment object
}
