% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{register_model_from_run}
\alias{register_model_from_run}
\title{Register a model for operationalization.}
\usage{
register_model_from_run(run, model_name, model_path = NULL,
  tags = NULL, properties = NULL, description = NULL,
  datasets = NULL, sample_input_dataset = NULL,
  sample_output_dataset = NULL, resource_configuration = NULL)
}
\arguments{
\item{run}{The \code{Run} object.}

\item{model_name}{The name of the model.}

\item{model_path}{The relative cloud path to the model, for example,
"outputs/modelname". When not specified, \code{model_name} is used as the path.}

\item{tags}{A dictionary of key value tags to assign to the model.}

\item{properties}{A dictionary of key value properties to assign to the model.
These properties cannot be changed after model creation, however new key-value pairs can be added.}

\item{description}{An optional description of the model.}

\item{datasets}{A list of tuples where the first element describes the dataset-model
relationship and the second element is the dataset.}

\item{sample_input_dataset}{Sample input dataset for the registered model.}

\item{sample_output_dataset}{Sample output dataset for the registered model.}

\item{resource_configuration}{`ResourceConfiguration`` object to run the registered model.}
}
\value{
The registered Model.
}
\description{
Register a model for operationalization.
}
\seealso{
\code{\link{resource_configuration}}
}
