% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-local.R
\name{reload_local_webservice_assets}
\alias{reload_local_webservice_assets}
\title{Reload the LocalWebservice's execution script and dependencies.
This restarts the service's container with copies of updated assets,
including the execution script and local dependencies, but it does not
rebuild the underlying image. Accordingly, changes to Conda/pip dependencies
or custom Docker steps will not be reflected in the reloaded LocalWebservice.
To handle those changes call LocalWebservice.update(), instead.}
\usage{
reload_local_webservice_assets(webservice, wait = FALSE)
}
\arguments{
\item{webservice}{LocalWebservice object.}

\item{wait}{Wait for the service's container to reach a healthy state.}
}
\description{
Reload the LocalWebservice's execution script and dependencies.
This restarts the service's container with copies of updated assets,
including the execution script and local dependencies, but it does not
rebuild the underlying image. Accordingly, changes to Conda/pip dependencies
or custom Docker steps will not be reflected in the reloaded LocalWebservice.
To handle those changes call LocalWebservice.update(), instead.
}
