% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{resource_configuration}
\alias{resource_configuration}
\title{Initialize the  ResourceConfiguration.}
\usage{
resource_configuration(cpu = NULL, memory_in_gb = NULL, gpu = NULL)
}
\arguments{
\item{cpu}{The number of CPU cores to allocate for this resource. Can be a decimal.}

\item{memory_in_gb}{The amount of memory (in GB) to allocate for this resource.
Can be a decimal If \code{TRUE}, decode the raw log bytes to a string.}

\item{gpu}{The number of GPUs to allocate for this resource.}
}
\value{
The \code{ResourceConfiguration} object.
}
\description{
Initialize the  ResourceConfiguration.
}
\examples{
\dontrun{
rc <- resource_configuration(2, 2, 0)

registered_model <- register_model_from_run(run, "my_model_name",
                                            "path_to_my_model",
                                            resource_configuration = rc)
}
}
\seealso{
\code{\link{register_model_from_run}}
}
