% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{save_model_package_files}
\alias{save_model_package_files}
\title{Save the package output to a local directory.
This can only be used with a Dockerfile package.}
\usage{
save_model_package_files(package, output_directory)
}
\arguments{
\item{package}{Package created with model(s) and dependencies.}

\item{output_directory}{Local directory that will be created to contain
the contents of the package.}
}
\description{
Save the package output to a local directory.
This can only be used with a Dockerfile package.
}
