% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{serialize_model}
\alias{serialize_model}
\title{Convert this Model into a json serialized dictionary}
\usage{
serialize_model(model)
}
\arguments{
\item{model}{The model to download}
}
\description{
Convert this Model into a json serialized dictionary
}
