% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{serialize_webservice}
\alias{serialize_webservice}
\title{Convert this Webservice into a json serialized dictionary.}
\usage{
serialize_webservice(webservice)
}
\arguments{
\item{webservice}{The webservice object.}
}
\value{
The json representation of this Webservice
}
\description{
Convert this Webservice into a json serialized dictionary.
}
