% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{set_datastore_to_download}
\alias{set_datastore_to_download}
\title{Set datastore to download mode}
\usage{
set_datastore_to_download(ds)
}
\arguments{
\item{ds}{datastore object}
}
\description{
Set datastore to download mode
}

