% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{set_datastore_to_mount}
\alias{set_datastore_to_mount}
\title{Set datastore to mount mode}
\usage{
set_datastore_to_mount(ds)
}
\arguments{
\item{ds}{datastore object}
}
\description{
Set datastore to mount mode
}

