% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{set_default_datastore}
\alias{set_default_datastore}
\title{Set the default datastore for the workspace.}
\usage{
set_default_datastore(workspace, datastore_name)
}
\arguments{
\item{workspace}{The workspace object}

\item{datastore_name}{The name of the datastore to be set as default}
}
\description{
Set the default datastore for the workspace.
}
