% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{set_secrets}
\alias{set_secrets}
\title{Add a named list of secrets eg. list("name" = "value") into the keyvault}
\usage{
set_secrets(keyvault, secrets)
}
\arguments{
\item{keyvault}{keyvault object}
}
\description{
Add a named list of secrets eg. list("name" = "value") into the keyvault
}
