% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{skip_from_dataset}
\alias{skip_from_dataset}
\title{Skip file streams from the top of the dataset by the specified count.}
\usage{
skip_from_dataset(dataset, count)
}
\arguments{
\item{dataset}{The Dataset object.}

\item{count}{The number of file streams to skip.}
}
\value{
A new Dataset object representing the dataset with file streams skipped.
}
\description{
Skip file streams from the top of the dataset by the specified count.
}
