% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_azureml_parallel.R
\name{split_tasks}
\alias{split_tasks}
\title{Splits the job into parallel tasks.}
\usage{
split_tasks(args_list, node_count, process_count_per_node)
}
\arguments{
\item{args_list}{The list of arguments which are distributed across all the
processes.}

\item{node_count}{Number of nodes in the AmlCompute cluster.}

\item{process_count_per_node}{Number of processes per node.}
}
\description{
Splits the job into parallel tasks.
}
