% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{start_logging_run}
\alias{start_logging_run}
\title{Start an interactive logging session in the specified experiment.}
\usage{
start_logging_run(experiment, outputs = NULL,
  snapshot_directory = NULL)
}
\arguments{
\item{experiment}{The \code{Experiment} object.}

\item{outputs}{Optional outputs directory to track.}

\item{snapshot_directory}{Optional directory to take snapshot of. Setting to
NULL will take no snapshot.}
}
\value{
Return a started run.
}
\description{
Create an interactive run that allows the user to log
metrics & artifacts to a run record in an experiment locally.

Any metrics that are logged during the session are added to the run
record in the experiment. If an output directory is specified, the contents
of that directory is uploaded as run artifacts upon run completion.
}
