% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{submit_child_run}
\alias{submit_child_run}
\title{Submit an experiment and return the active child run}
\usage{
submit_child_run(parent_run, config = NULL, tags = NULL)
}
\arguments{
\item{parent_run}{The \code{HyperDriveRun} object.}

\item{config}{The config to be submitted, HyperDriveConfig or estimator object}

\item{tags}{Tags to be added to the submitted run, e.g., {"tag": "value"}.}
}
\value{
A run object.
}
\description{
Submit an experiment and return the active child run.
}
