% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{submit_experiment}
\alias{submit_experiment}
\title{Submit experiment}
\usage{
submit_experiment(config, experiment, tags = NULL)
}
\arguments{
\item{config}{runconfig or estimator}

\item{experiment}{experiment object}

\item{tags}{Tags to be added to the submitted run. A named list eg. 
list("tag" = "value")}
}
\value{
run object
}
\description{
Submit experiment
}
