% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{take_from_dataset}
\alias{take_from_dataset}
\title{Take a sample of file streams from top of the dataset by the specified count.}
\usage{
take_from_dataset(dataset, count)
}
\arguments{
\item{dataset}{The Dataset object.}

\item{count}{The number of file streams to take.}
}
\value{
A new Dataset object representing the sampled dataset.
}
\description{
Take a sample of file streams from top of the dataset by the specified count.
}
