% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{take_sample_from_dataset}
\alias{take_sample_from_dataset}
\title{Take a random sample of file streams in the dataset approximately by the probability specified.}
\usage{
take_sample_from_dataset(dataset, probability, seed = NULL)
}
\arguments{
\item{dataset}{The Dataset object.}

\item{probability}{The probability of a file stream being included in the sample.}

\item{seed}{An optional seed to use for the random generator.}
}
\value{
A new Dataset object representing the sampled dataset.
}
\description{
Take a random sample of file streams in the dataset approximately by the probability specified.
}
