% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{truncation_selection_policy}
\alias{truncation_selection_policy}
\title{Create Truncation Selection policy for HyperDrive runs}
\usage{
truncation_selection_policy(truncation_percentage,
  evaluation_interval = 1L, delay_evaluation = 0L)
}
\arguments{
\item{truncation_percentage}{percentage of lowest performing runs to
terminate at each interval}

\item{evaluation_interval}{frequency for applying policy}

\item{delay_evaluation}{how many intervals to delay the first evaluation}
}
\value{
EarlyTerminationPolicy object
}
\description{
Create Truncation Selection policy for HyperDrive runs
}
