% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{uniform}
\alias{uniform}
\title{Specify a uniform distribution from which samples are taken}
\usage{
uniform(min_value, max_value)
}
\arguments{
\item{min_value}{minimum value in the range (inclusive)}

\item{max_value}{maximum value in the range (inclusive)}
}
\value{
the stochastic expression
}
\description{
Specify a uniform distribution from which samples are taken
}
