% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{unregister_all_dataset_versions}
\alias{unregister_all_dataset_versions}
\title{Unregister all versions under the registration name of this dataset from the workspace.}
\usage{
unregister_all_dataset_versions(dataset)
}
\arguments{
\item{dataset}{The dataset to be unregistered.}
}
\value{
None
}
\description{
Unregister all versions under the registration name of this dataset from the workspace.
}
