% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{unregister_datastore}
\alias{unregister_datastore}
\title{Unregisters the datastore. the underlying storage service will not 
be deleted.}
\usage{
unregister_datastore(datastore)
}
\arguments{
\item{datastore}{datastore object}
}
\description{
Unregisters the datastore. the underlying storage service will not 
be deleted.
}
