% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-aci.R
\name{update_aci_webservice}
\alias{update_aci_webservice}
\title{Update the Webservice with provided properties.
Values left as None will remain unchanged in this Webservice.}
\usage{
update_aci_webservice(webservice, tags = NULL, properties = NULL,
  description = NULL, auth_enabled = NULL, ssl_enabled = NULL,
  ssl_cert_pem_file = NULL, ssl_key_pem_file = NULL,
  ssl_cname = NULL, enable_app_insights = NULL, models = NULL,
  inference_config = NULL)
}
\arguments{
\item{webservice}{AciWebservice object.}

\item{tags}{Dictionary of key value tags to give this Webservice. 
Will replace existing tags.}

\item{properties}{Dictionary of key value properties to add to existing 
properties dictionary.}

\item{description}{A description to give this Webservice.}

\item{auth_enabled}{Enable or disable auth for this Webservice.}

\item{ssl_enabled}{Whether or not to enable SSL for this Webservice.}

\item{ssl_cert_pem_file}{The cert file needed if SSL is enabled.}

\item{ssl_key_pem_file}{The key file needed if SSL is enabled.}

\item{ssl_cname}{The cname for if SSL is enabled.}

\item{enable_app_insights}{Whether or not to enable AppInsights for this 
Webservice.}

\item{models}{A list of Model objects to package into the updated service.}

\item{inference_config}{An InferenceConfig object used to provide the 
required model deployment properties.}
}
\description{
Update the Webservice with provided properties.
Values left as None will remain unchanged in this Webservice.
}
