% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-local.R
\name{update_local_webservice}
\alias{update_local_webservice}
\title{Update the LocalWebservice with provided properties.
Values left as None will remain unchanged in this LocalWebservice.}
\usage{
update_local_webservice(webservice, models = NULL,
  deployment_config = NULL, wait = FALSE, inference_config = NULL)
}
\arguments{
\item{webservice}{LocalWebservice object.}

\item{models}{A new list of models contained in the LocalWebservice.}

\item{deployment_config}{Deployment configuration options to apply to the 
LocalWebservice.}

\item{wait}{Wait for the service's container to reach a healthy state.}

\item{inference_config}{An InferenceConfig object used to provide the 
required model deployment properties.}
}
\description{
Update the LocalWebservice with provided properties.
Values left as None will remain unchanged in this LocalWebservice.
}
