% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{upload_files_to_datastore}
\alias{upload_files_to_datastore}
\title{Upload the data from the local file system to the Azure storage this
datastore points to.}
\usage{
upload_files_to_datastore(datastore, files, relative_root = NULL,
  target_path = NULL, overwrite = FALSE, show_progress = TRUE)
}
\arguments{
\item{datastore}{datastore object}

\item{files}{list of absolute path to files to upload}

\item{relative_root}{the base path from which is used to determine the path
of the files in the file share. For example, if we upload /path/to/file.txt,
and we define base path to be /path, when file.txt is uploaded to the file
share, it will have the path of /to/file.txt. If target_path is also given,
then it will be used as the prefix for the derived path from above. The base
path must be a common path of all of the files, otherwise an exception will
be thrown, defaults to None, which will find the common path.}

\item{target_path}{location in the file share to upload the data to, defaults
to None, the root}

\item{overwrite}{overwrites, defaults to FALSE}

\item{show_progress}{progress of upload in the console, defaults to TRUE}
}
\description{
Upload the data from the local file system to the Azure storage this
datastore points to.
}
