% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{upload_files_to_run}
\alias{upload_files_to_run}
\title{Upload files to the run record.}
\usage{
upload_files_to_run(names, paths, timeout_seconds = NULL, run = NULL)
}
\arguments{
\item{names}{A character vector of file names to upload. If set,
paths must also be set.}

\item{paths}{A character vector of relative local paths to the files to
upload. If set, names is required.}

\item{timeout_seconds}{The timeout for uploading files.}

\item{run}{Run object.}
}
\description{
Upload files to the run record.
}
