% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{upload_to_datastore}
\alias{upload_to_datastore}
\title{Upload the data from the local file system to the Azure storage this datastore points to.}
\usage{
upload_to_datastore(datastore, src_dir, target_path = NULL,
  overwrite = FALSE, show_progress = TRUE)
}
\arguments{
\item{datastore}{datastore object}

\item{src_dir}{the local directory to upload}

\item{target_path}{location in the file share to upload the data to, defaults to None, the root}

\item{overwrite}{overwrites, defaults to FALSE}

\item{show_progress}{progress of upload in the console, defaults to TRUE}
}
\description{
Upload the data from the local file system to the Azure storage this datastore points to.
}
