% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{view_run_details}
\alias{view_run_details}
\title{Plot table of run details in Viewer}
\usage{
view_run_details(run)
}
\arguments{
\item{run}{run used for plotting}
}
\description{
Plot table of run details in Viewer
}
