% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{wait_for_compute}
\alias{wait_for_compute}
\title{Wait for cluster's instantiation}
\usage{
wait_for_compute(cluster, show_output = TRUE)
}
\arguments{
\item{cluster}{cluster object}

\item{show_output}{show output on console}
}
\description{
Wait for cluster's instantiation
}
