% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{wait_for_model_package_creation}
\alias{wait_for_model_package_creation}
\title{Wait for the package to finish creating.}
\usage{
wait_for_model_package_creation(package, show_output = FALSE)
}
\arguments{
\item{package}{Package created with model(s) and dependencies.}

\item{show_output}{Boolean option to print more verbose output. Defaults to
FALSE.}
}
\description{
Wait for the package to finish creating.
}
