% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{wait_for_run_completion}
\alias{wait_for_run_completion}
\title{Wait for the completion of this run}
\usage{
wait_for_run_completion(run, show_output = TRUE)
}
\arguments{
\item{run}{Run object}

\item{show_output}{print verbose output to console}
}
\description{
Wait for the completion of this run
}
