% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{write_workspace_config}
\alias{write_workspace_config}
\title{Write out the Workspace ARM properties to a config file}
\usage{
write_workspace_config(workspace, path = NULL, file_name = NULL)
}
\arguments{
\item{workspace}{The workspace whose config has to be written down.}

\item{path}{User provided location to write the config.json file.
The parameter defaults to the current working directory.}

\item{file_name}{Name to use for the config file. The parameter defaults to
config.json.}
}
\description{
Write out the Workspace ARM properties to a config file
}
