% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{attach_aks_compute}
\alias{attach_aks_compute}
\title{Attach an already existing AKS compute resource with the provided workspace}
\usage{
attach_aks_compute(workspace, cluster_name, resource_id,
  resource_group = NULL)
}
\arguments{
\item{workspace}{The workspace object to attach the Compute object to.}

\item{cluster_name}{The AKS cluster name}

\item{resource_id}{The Azure resource ID for the compute resource being attached}

\item{resource_group}{Name of the resource group in which the AKS is located.}
}
\value{
An AksCompute object representation of the compute object
}
\description{
Attach an already existing AKS compute resource with the provided workspace
}
