% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modules.R
\docType{data}
\name{azureml}
\alias{azureml}
\title{azureml module
User can access functions/modules in azureml that are not exposed through the
exported R functions.}
\format{
An object of class \code{python.builtin.module} (inherits from \code{python.builtin.object}) of length 5.
}
\usage{
azureml
}
\description{
azureml module
User can access functions/modules in azureml that are not exposed through the
exported R functions.
}
\keyword{datasets}
