% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{choice}
\alias{choice}
\title{Specify a discrete set of options to sample from}
\usage{
choice(options)
}
\arguments{
\item{options}{A vector of values to choose from.}
}
\value{
A list of the stochastic expression.
}
\description{
Specify a discrete set of options to sample the hyperparameters
from.
}
\seealso{
\code{random_parameter_sampling()}, \code{grid_parameter_sampling()},
\code{bayesian_parameter_sampling()}
}
