% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{complete_run}
\alias{complete_run}
\title{Mark a run as completed.}
\usage{
complete_run(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\value{
None
}
\description{
Mark the run as completed. Use for an interactive logging run.
}
\seealso{
\code{\link[=start_logging_run]{start_logging_run()}}
}
