% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{container_registry}
\alias{container_registry}
\title{Specify Azure Container Registry details}
\usage{
container_registry(address = NULL, username = NULL, password = NULL)
}
\arguments{
\item{address}{A string of the DNS name or IP address of the
Azure Container Registry (ACR).}

\item{username}{A string of the username for ACR.}

\item{password}{A string of the password for ACR.}
}
\value{
The \code{ContainerRegistry} object.
}
\description{
Returns a \code{ContainerRegistry} object with the details for an
Azure Container Registry (ACR). This is needed when a custom
Docker image used for training or deployment is located in
a private image registry. Provide a \code{ContainerRegistry} object
to the \code{image_registry_details} parameter of either \code{r_environment()}
or \code{estimator()}.
}
\seealso{
\code{r_environment()}, \code{estimator()}
}
