% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{create_child_run}
\alias{create_child_run}
\title{Create a child run}
\usage{
create_child_run(parent_run, name = NULL, run_id = NULL, outputs = NULL)
}
\arguments{
\item{parent_run}{The parent \code{Run} object.}

\item{name}{An optional name for the child run, typically specified for a "part"}

\item{run_id}{An optional run ID for the child, otherwise it is auto-generated.
Typically this parameter is not set.}

\item{outputs}{Optional outputs directory to track for the child.}
}
\value{
The child run, a \code{Run} object.
}
\description{
Create a child run. This is used to isolate part of a run into a subsection.
}
