% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{create_hyperdrive_config}
\alias{create_hyperdrive_config}
\title{Create configuration for a HyperDrive run}
\usage{
create_hyperdrive_config(hyperparameter_sampling, primary_metric_name,
  primary_metric_goal, max_total_runs, max_concurrent_runs = NULL,
  max_duration_minutes = 10080, policy = NULL, estimator = NULL)
}
\arguments{
\item{hyperparameter_sampling}{hyperparameter sampling option}

\item{primary_metric_name}{name of primary metric}

\item{primary_metric_goal}{goal of primary metric}

\item{max_total_runs}{maximum number of runs to queue}

\item{max_concurrent_runs}{maximum number of runs to start concurrently}

\item{max_duration_minutes}{maximum time to allow runs}

\item{policy}{termination policy}

\item{estimator}{estimator object (N/A if using run_config or pipeline)}
}
\value{
HyperDrive config object
}
\description{
Create configuration for a HyperDrive run
}
