% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-local.R
\name{delete_local_webservice}
\alias{delete_local_webservice}
\title{Delete a local web service from the local machine}
\usage{
delete_local_webservice(webservice, delete_cache = TRUE, delete_image = FALSE)
}
\arguments{
\item{webservice}{The \code{LocalWebservice} object.}

\item{delete_cache}{If \code{TRUE}, delete the temporary files cached for
the service.}

\item{delete_image}{If \code{TRUE}, delete the service's Docker image.}
}
\value{
None
}
\description{
Delete a local web service from the local machine. This function call
is not asynchronous; it runs until the service is deleted.
}
