% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{delete_model}
\alias{delete_model}
\title{Delete a model from its associated workspace}
\usage{
delete_model(model)
}
\arguments{
\item{model}{The \code{Model} object.}
}
\value{
None
}
\description{
Delete the registered model from its associated workspace. Note that
you cannot delete a registered model that is being used by an active
web service deployment.
}
