% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{delete_secrets}
\alias{delete_secrets}
\title{Delete secrets from a keyvault}
\usage{
delete_secrets(keyvault, secrets)
}
\arguments{
\item{keyvault}{The \code{Keyvault} object.}

\item{secrets}{A vector of secret names.}
}
\value{
None
}
\description{
Delete secrets from the keyvault associated with the workspace for
a specified set of secret names.
}
