% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{delete_webservice}
\alias{delete_webservice}
\title{Delete a web service from a given workspace}
\usage{
delete_webservice(webservice)
}
\arguments{
\item{webservice}{The \code{AciWebservice} or \code{AksWebservice} object.}
}
\value{
None
}
\description{
Delete a deployed ACI or AKS web service from the given workspace.
This function call is not asynchronous; it runs until the resource is
deleted.

To delete a \code{LocalWebservice} see \code{delete_local_webservice()}.
}
