% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{detach_aks_compute}
\alias{detach_aks_compute}
\title{Detach an AksCompute cluster from its associated workspace}
\usage{
detach_aks_compute(cluster)
}
\arguments{
\item{cluster}{The \code{AksCompute} object.}
}
\value{
None
}
\description{
Detach the AksCompute cluster from its associated workspace. No
underlying cloud resource will be deleted; the association will
just be removed.
}
