% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{estimator}
\alias{estimator}
\title{Create Estimator}
\usage{
estimator(source_directory, compute_target = NULL, vm_size = NULL,
  vm_priority = NULL, entry_script = NULL, script_params = NULL,
  cran_packages = NULL, github_packages = NULL,
  custom_url_packages = NULL, custom_docker_image = NULL,
  image_registry_details = NULL, use_gpu = FALSE,
  environment_variables = NULL, shm_size = NULL,
  max_run_duration_seconds = NULL, environment = NULL)
}
\arguments{
\item{source_directory}{A local directory containing experiment configuration
files.}

\item{compute_target}{The ComputeTarget where training will happen.
Currently only support AmlCompute type.}

\item{vm_size}{The VM size of the compute target that will be created for the
training. Supported values: Any Azure VM size. The list of available VM sizes
are listed here:
https://docs.microsoft.com/azure/cloud-services/cloud-services-sizes-specs}

\item{vm_priority}{The VM priority of the compute target that will be created
for the training. If not specified, it will be defaulted to 'dedicated'.
Supported values: 'dedicated' and 'lowpriority'. This takes effect only when
the vm_size param is specified in the input.}

\item{entry_script}{A string representing the relative path to the file used
to start training.}

\item{script_params}{A named list containing parameters to the entry_script.}

\item{cran_packages}{character vector of cran packages to be installed.}

\item{github_packages}{character vector of github packages to be installed.}

\item{custom_url_packages}{character vector of packages to be installed from
local, directory or custom url.}

\item{custom_docker_image}{The name of the docker image from which the image
to use for training will be built. If not set, a default CPU based image will
be used as the base image.}

\item{image_registry_details}{The details of the Docker image registry.}

\item{use_gpu}{Indicates whether the environment to run the experiment should
support GPUs.
If TRUE, a GPU - based default Docker image will be used in the environment.
If FALSE, a CPU - based image will be used. Default Docker images
(CPU orGPU) will be used only if the 'custom_docker_image' parameter is not
set.}

\item{environment_variables}{A dictionary of environment variables names 
and values. These environment variables are set on the process where user 
script is being executed.}

\item{shm_size}{The size of the Docker container's shared memory block.}

\item{max_run_duration_seconds}{The maximum allowed time for the run. 
Azure ML will attempt to automatically cancel the run if it take longer 
than this value.}

\item{environment}{the R environment where the experiment is executed.
Mutually exclusive with custom_docker_image, image_registry_details, use_gpu,
environment_variables, shm_size, max_run_duration_seconds, cran_packages,
github_packages and custom_url_packages.}
}
\description{
Create Estimator
}
