% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{estimator}
\alias{estimator}
\title{Create Estimator}
\usage{
estimator(source_directory, compute_target = NULL, vm_size = NULL,
  vm_priority = NULL, entry_script = NULL, script_params = NULL,
  use_docker = TRUE, cran_packages = NULL, github_packages = NULL,
  custom_url_packages = NULL, custom_docker_image = NULL,
  inputs = NULL)
}
\arguments{
\item{source_directory}{A local directory containing experiment configuration files.}

\item{compute_target}{The ComputeTarget where training will happen. This can either be an object or the
string "local".}

\item{vm_size}{The VM size of the compute target that will be created for the training. Supported values:
Any Azure VM size. The list of available VM sizes are listed here:
https://docs.microsoft.com/en-us/azure/cloud-services/cloud-services-sizes-specs}

\item{vm_priority}{The VM priority of the compute target that will be created for the training. If not specified,
it will be defaulted to 'dedicated'. Supported values: 'dedicated' and 'lowpriority'. This takes effect only when
the vm_size param is specified in the input.}

\item{entry_script}{A string representing the relative path to the file used to start training.}

\item{script_params}{A named list containing parameters to the entry_script.}

\item{use_docker}{A bool value indicating if the environment to run the experiment should be docker-based.}

\item{cran_packages}{character vector of cran packages to be installed.}

\item{github_packages}{character vector of github packages to be installed.}

\item{custom_url_packages}{character vector of packages to be installed from local, directory or custom url.}

\item{custom_docker_image}{The name of the docker image from which the image to use for training will be built. If
not set, a default CPU based image will be used as the base image.}

\item{inputs}{list of data references as input}
}
\description{
Create Estimator
}
