% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Create an Azure Machine Learning experiment}
\usage{
experiment(workspace, name)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{name}{A string of the experiment name. The name must be between
3-36 characters, start with a letter or number, and can only contain
letters, numbers, underscores, and dashes.}
}
\value{
The \code{Experiment} object.
}
\description{
An experiment is a grouping of many runs from a specified script.
}
\examples{
\dontrun{
ws <- load_workspace_from_config()
exp <- experiment(ws, name = 'myexperiment')
}
}
\seealso{
\code{submit_experiment()}
}
