% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{generate_new_webservice_key}
\alias{generate_new_webservice_key}
\title{Regenerate one of a web service's keys}
\usage{
generate_new_webservice_key(webservice, key_type)
}
\arguments{
\item{webservice}{The \code{AciWebservice} or \code{AksWebservice} object.}

\item{key_type}{A string of which key to regenerate. Options are
"Primary" or "Secondary".}
}
\value{
None
}
\description{
Regenerate either the primary or secondary authentication key for
an \code{AciWebservice} or \code{AksWebservice}.The web service must have
been deployed with key-based authentication enabled.

Not supported for \code{LocalWebservice} deployments.
}
