% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{get_aks_compute_credentials}
\alias{get_aks_compute_credentials}
\title{Retrieve the credentials for the AKS target}
\usage{
get_aks_compute_credentials(cluster)
}
\arguments{
\item{cluster}{cluster object}
}
\value{
Credentials for the AKS target
}
\description{
Retrieve the credentials for the AKS target
}
