% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{get_child_run_hyperparameters}
\alias{get_child_run_hyperparameters}
\title{Get the hyperparameters for all child runs}
\usage{
get_child_run_hyperparameters(hyperdrive_run)
}
\arguments{
\item{hyperdrive_run}{The \code{HyperDriveRun} object.}
}
\value{
The named list of hyperparameters where element name
is the run_id, e.g. \code{list("run_id" = hyperparameters)}.
}
\description{
Return the hyperparameters for all the child runs of the
HyperDrive run.
}
