% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_child_runs}
\alias{get_child_runs}
\title{Get all children for the current run selected by specified filters}
\usage{
get_child_runs(
  parent_run,
  recursive = FALSE,
  tags = NULL,
  properties = NULL,
  type = NULL,
  status = NULL
)
}
\arguments{
\item{parent_run}{The parent \code{Run} object.}

\item{recursive}{Boolean indicating whether to recurse through all descendants.}

\item{tags}{If specified, returns runs matching specified "tag" or list(tag = value).}

\item{properties}{If specified, returns runs matching specified "property" or list(property = value).}

\item{type}{If specified, returns runs matching this type.}

\item{status}{If specified, returns runs with status specified "status".}
}
\value{
A list of child runs, \code{Run} objects.
}
\description{
Get all children for the current run selected by specified filters.
}
