% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{get_datastore}
\alias{get_datastore}
\title{Get a datastore by name}
\usage{
get_datastore(workspace, datastore_name)
}
\arguments{
\item{workspace}{The workspace object}

\item{datastore_name}{The name of the datastore.}
}
\value{
The corresponding datastore for that name.
}
\description{
Get a datastore by name
}
