% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{get_datastore}
\alias{get_datastore}
\title{Get an existing datastore}
\usage{
get_datastore(workspace, datastore_name)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{datastore_name}{A string of the name of the datastore.}
}
\value{
The \code{azureml.data.azure_sql_database.AzureBlobDatastore},
\code{azureml.data.azure_sql_database.AzureFileDatastore},
\code{azureml.data.azure_sql_database.AzureSqlDatabaseDatastore},
\code{azureml.data.azure_data_lake_datastore.AzureDataLakeGen2Datastore},
\code{azureml.data.azure_postgre_sql_datastore.AzurePostgreSqlDatastore}, or
\code{azureml.data.azure_sql_database.AzureSqlDatabaseDatastore} object.
}
\description{
Get the corresponding datastore object for an existing
datastore by name from the given workspace.
}
