% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_default_datastore}
\alias{get_default_datastore}
\title{Get the default datastore for a workspace}
\usage{
get_default_datastore(workspace)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}
}
\value{
The default \code{Datastore} object.
}
\description{
Returns the default datastore associated with the workspace.
}
