% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{get_default_keyvault}
\alias{get_default_keyvault}
\title{Get the default keyvault for a workspace}
\usage{
get_default_keyvault(workspace)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}
}
\value{
The \code{Keyvault} object.
}
\description{
Returns a \code{Keyvault} object representing the default
\href{https://docs.microsoft.com/en-us/azure/key-vault/key-vault-overview}{Azure Key Vault}
associated with the workspace.
}
\seealso{
\code{\link[=set_secrets]{set_secrets()}} \code{\link[=get_secrets]{get_secrets()}} \code{\link[=list_secrets]{list_secrets()}} \code{\link[=delete_secrets]{delete_secrets()}}
}
