% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/environment.R
\name{get_environment}
\alias{get_environment}
\title{Get an existing environment}
\usage{
get_environment(workspace, name, version = NULL)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{name}{A string of the name of the environment.}

\item{version}{A string of the version of the environment.}
}
\value{
The \code{Environment} object.
}
\description{
Returns an \code{Environment} object for an existing environment in
the workspace.
}
\examples{
\dontrun{
ws <- load_workspace_from_config()
env <- get_environment(ws, name = 'myenv', version = '1')
}
}
