% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{get_input_dataset_from_run}
\alias{get_input_dataset_from_run}
\title{Return the named list for input datasets.}
\usage{
get_input_dataset_from_run(name, run = NULL)
}
\arguments{
\item{name}{The name of the input dataset}

\item{run}{The run taking the dataset as input}
}
\value{
A dataset object corresponding to the "name"
}
\description{
Return the named list for input datasets.
}
