% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{get_model_package_creation_logs}
\alias{get_model_package_creation_logs}
\title{Get the model package creation logs}
\usage{
get_model_package_creation_logs(package, decode = TRUE, offset = 0)
}
\arguments{
\item{package}{The \code{ModelPackage} object.}

\item{decode}{If \code{TRUE}, decode the raw log bytes to a string.}

\item{offset}{An int of the byte offset from which to start
reading the logs.}
}
\value{
A string or character vector of package creation logs.
}
\description{
Retrieve the creation logs from packaging a model with
\code{package_model()}.
}
