% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run}
\alias{get_run}
\title{Get an experiment run}
\usage{
get_run(experiment, run_id)
}
\arguments{
\item{experiment}{The \code{Experiment} object.}

\item{run_id}{A string of the run ID for the run.}
}
\value{
The \code{Run} object.
}
\description{
Given the associated experiment and run ID, return the
run object for a previously submitted/tracked run.
}
