% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run_details_with_logs}
\alias{get_run_details_with_logs}
\title{Get the details of a run along with the log files' contents}
\usage{
get_run_details_with_logs(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\value{
A named list of the run details and log file contents.
}
\description{
Get the details of a run along with the log files' contents
}
\seealso{
\code{\link[=get_run_details]{get_run_details()}}
}
