% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run_file_names}
\alias{get_run_file_names}
\title{List the files that are stored in association with a run}
\usage{
get_run_file_names(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\value{
A list of strings of the paths for existing artifacts
in the run record.
}
\description{
Get the list of files stored in a run record.
}
\seealso{
\code{\link[=download_file_from_run]{download_file_from_run()}} \code{\link[=download_files_from_run]{download_files_from_run()}}
}
