% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{get_run_metrics}
\alias{get_run_metrics}
\title{Get the metrics for run}
\usage{
get_run_metrics(run)
}
\arguments{
\item{run}{Run object}
}
\value{
named list containing metrics associated with the run.
}
\description{
Get the metrics for run
}
