% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{get_secrets}
\alias{get_secrets}
\title{Get secrets from a keyvault}
\usage{
get_secrets(keyvault, secrets)
}
\arguments{
\item{keyvault}{The \code{Keyvault} object.}

\item{secrets}{A vector of secret names.}
}
\value{
A named list of found and not found secrets, where element
name corresponds to the secret name. If a secret was not found, the
corresponding element will be \code{NULL}.
}
\description{
Returns the secret values from the keyvault associated with the
workspace for a given set of secret names. For runs submitted using
\code{submit_experiment()}, you can use \code{get_secrets_from_run()} instead,
as that method shortcuts workspace instantiation (since a submitted
run is aware of its workspace).
}
