% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{get_webservice}
\alias{get_webservice}
\title{Get a deployed web service}
\usage{
get_webservice(workspace, name)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{name}{A string of the name of the web service to retrieve.}
}
\value{
The \code{LocalWebservice}, \code{AciWebservice}, or \code{AksWebservice} object.
}
\description{
Return the corresponding Webservice object of a deployed web service from
a given workspace.
}
