% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install.R
\name{install_azureml}
\alias{install_azureml}
\title{Install azureml sdk package}
\usage{
install_azureml(version = NULL, environment = "r-azureml",
  conda_python_version = "3.6")
}
\arguments{
\item{version}{pip version}

\item{environment}{name of environment to create}

\item{conda_python_version}{version of python for conda environment}
}
\description{
Install azureml sdk package
}
