% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keyvault.R
\name{list_secrets}
\alias{list_secrets}
\title{List the secrets in a keyvault}
\usage{
list_secrets(keyvault)
}
\arguments{
\item{keyvault}{The \code{Keyvault} object.}
}
\value{
A list of secret names.
}
\description{
Returns the list of secret names for all the secrets in the keyvault
associated with the workspace.
}
