% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{list_supported_vm_sizes}
\alias{list_supported_vm_sizes}
\title{List the supported VM sizes in a region}
\usage{
list_supported_vm_sizes(workspace, location = NULL)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{location}{A string of the location of the cluster. If not specified,
will default to the workspace location.}
}
\value{
A data frame of supported VM sizes in a region with name of the VM, VCPUs,
RAM.
}
\description{
List the supported VM sizes in a region
}
