% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{load_workspace_from_config}
\alias{load_workspace_from_config}
\title{Load workspace from config}
\usage{
load_workspace_from_config(path)
}
\arguments{
\item{path}{Path to the config file or starting directory for search.
The parameter defaults to starting the search in the current directory.}
}
\description{
Load workspace from config
}
