% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice-local.R
\name{local_webservice_deployment_config}
\alias{local_webservice_deployment_config}
\title{Create a deployment config for deploying a local web service}
\usage{
local_webservice_deployment_config(port = NULL)
}
\arguments{
\item{port}{An int of the local port on which to expose the service's
HTTP endpoint.}
}
\value{
The \code{LocalWebserviceDeploymentConfiguration} object.
}
\description{
You can deploy a model locally for limited testing and troubleshooting.
To do so, you will need to have Docker installed on your local machine.

If you are using an Azure Machine Learning Compute Instance for
development, you can also deploy locally on your compute instance.
}
\examples{
\dontrun{
deployment_config <- local_webservice_deployment_config(port = 8890)
}
}
