% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_confusion_matrix_to_run}
\alias{log_confusion_matrix_to_run}
\title{Log a confusion matrix metric to a run}
\usage{
log_confusion_matrix_to_run(name, value, description = "", run = NULL)
}
\arguments{
\item{name}{A string of the name of the metric.}

\item{value}{A named list containing name, version, and data properties.}

\item{description}{(Optional) A string of the metric description.}

\item{run}{The \code{Run} object. If not specified, will default
to the current run from the service context.}
}
\value{
None
}
\description{
Log a confusion matrix metric to a run
}
