% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{log_image_to_run}
\alias{log_image_to_run}
\title{Log an image metric to a run}
\usage{
log_image_to_run(name, path = NULL, plot = NULL, description = "", run = NULL)
}
\arguments{
\item{name}{A string of the name of the metric.}

\item{path}{A string of the path or stream of the image.}

\item{plot}{The ggplot2 plot to log as an image.}

\item{description}{(Optional) A string of the metric description.}

\item{run}{The \code{Run} object. If not specified, will default
to the current run from the service context.}
}
\value{
None
}
\description{
Log an image to the run with the give metric name. Use
\code{log_image_to_run()} to log an image file or ggplot2 plot to the
run. These images will be visible and comparable in the run
record.
}
