% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive_runconfig.R
\name{median_stopping_policy}
\alias{median_stopping_policy}
\title{Create Median Stopping policy for HyperDrive runs}
\usage{
median_stopping_policy(evaluation_interval = 1L, delay_evaluation = 0L)
}
\arguments{
\item{evaluation_interval}{frequency for applying policy}

\item{delay_evaluation}{how many intervals to delay the first evaluation}
}
\value{
EarlyTerminationPolicy object
}
\description{
Create Median Stopping policy for HyperDrive runs
}
