% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimator.R
\name{parallel_run_config}
\alias{parallel_run_config}
\title{Create a parallel run config}
\usage{
parallel_run_config(source_directory = ".", name = NULL,
  inputs = NULL, output = NULL, models = list(),
  compute_target = NULL, entry_script = NULL, arguments = lists(),
  mini_batch_size = 5, output_action = "append_row", node_count = 1,
  run_invocation_timeout = 600L, error_threshold = 10L,
  environment = NULL, allow_reuse = TRUE)
}
\description{
Create a parallel run config
}
