% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{plot_run_details}
\alias{plot_run_details}
\title{Generate table of run details}
\usage{
plot_run_details(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\value{
Datatable with run details
}
\description{
Generate table of run details including Run Status, Start Time,
Duration, Script Name, Arguments, link to web portal, and any run errors.
}
