% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{plot_run_details}
\alias{plot_run_details}
\title{Generate table of run details}
\usage{
plot_run_details(run)
}
\arguments{
\item{run}{The \code{Run} object.}
}
\value{
Datatable containing run details
}
\description{
Plot a table of run details including
\itemize{
\item ID
\item Status
\item Start Time
\item Duration
\item Script Name
\item Arguments
\item Link to Web Portal view
\item Errors
}
}
