% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hyperdrive.R
\name{primary_metric_goal}
\alias{primary_metric_goal}
\title{Define supported metric goals for hyperparameter tuning}
\usage{
primary_metric_goal(goal)
}
\arguments{
\item{goal}{A string of the metric goal (either "MAXIMIZE" or "MINIMIZE").}
}
\value{
The \code{PrimaryMetricGoal} object.
}
\description{
A metric goal is used to determine whether a higher value for a metric
is better or worse. Metric goals are used when comparing runs based on
the primary metric. For example, you may want to maximize accuracy or
minimize error.

The primary metric name and goal are specified to \code{hyperdrive_config()}
when you configure a HyperDrive run.
}
