% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model.R
\name{pull_model_package_image}
\alias{pull_model_package_image}
\title{Pull the Docker image from a \code{ModelPackage} to your local
Docker environment}
\usage{
pull_model_package_image(package)
}
\arguments{
\item{package}{The \code{ModelPackage} object.}
}
\value{
None
}
\description{
Pull the Docker image from a created \code{ModelPackage} to your
local Docker environment. The output of this call will
display the name of the image. For example:
\verb{Status: Downloaded newer image for myworkspacef78fd10.azurecr.io/package:20190822181338}.

This can only be used with a Docker image \code{ModelPackage} (where
\code{package_model()} was called with \code{generate_dockerfile = FALSE}).

After you've pulled the image, you can start a local container based
on this image using Docker commands.
}
\seealso{
\code{package_model()}
}
