% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/workspace.R
\name{set_default_datastore}
\alias{set_default_datastore}
\title{Set the default datastore for a workspace}
\usage{
set_default_datastore(workspace, datastore_name)
}
\arguments{
\item{workspace}{The \code{Workspace} object.}

\item{datastore_name}{The name of the datastore to be set as default.}
}
\value{
None
}
\description{
Set the default datastore associated with the workspace.
}
\seealso{
\code{\link[=get_default_datastore]{get_default_datastore()}}
}
