% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{submit_child_run}
\alias{submit_child_run}
\title{Submit an experiment and return the active child run}
\usage{
submit_child_run(parent_run, config = NULL, tags = NULL)
}
\arguments{
\item{parent_run}{The parent \code{Run} object.}

\item{config}{The \code{RunConfig} object}

\item{tags}{Tags to be added to the submitted run, e.g., {"tag": "value"}.}
}
\value{
A \code{Run} object.
}
\description{
Submit an experiment and return the active child run.
}
