% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{submit_experiment}
\alias{submit_experiment}
\title{Submit an experiment and return the active created run}
\usage{
submit_experiment(experiment, config, tags = NULL)
}
\arguments{
\item{experiment}{experiment object}

\item{config}{runconfig or estimator}

\item{tags}{Tags to be added to the submitted run. A named list
eg. list("tag" = "value")}
}
\value{
run object
}
\description{
Submit an experiment and return the active created run
}
