% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datastore.R
\name{unregister_datastore}
\alias{unregister_datastore}
\title{Unregister a datastore from its associated workspace}
\usage{
unregister_datastore(datastore)
}
\arguments{
\item{datastore}{The \code{AzureBlobDatastore} or \code{AzureFileDatastore} object.}
}
\value{
None
}
\description{
Unregister the datastore from its associated workspace. The
underlying Azure storage will not be deleted.
}
