% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{upload_files_to_run}
\alias{upload_files_to_run}
\title{Upload files to the run record.}
\usage{
upload_files_to_run(run, names, paths, timeout_seconds = NULL)
}
\arguments{
\item{run}{Run object.}

\item{names}{The names of the files to upload. If set, paths must also be 
set.}

\item{paths}{The relative local paths to the files to upload. If set, names 
is required.}

\item{timeout_seconds}{The timeout for uploading files.}
}
\description{
Upload files to the run record.
}
