% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{view_run_details}
\alias{view_run_details}
\title{Initialize run details widget}
\usage{
view_run_details(run)
}
\arguments{
\item{run}{Run object}
}
\description{
Plot table of run details in RStudio Viewer or browser.
If in RStudio, plot will update table values regularly. If not using
RStudio, the plot will be static. For more details than provided in the
plot, click the web view link.

If you are running this method from an RMarkdown file, the
run details table will show up in the code chunk output
instead of the Viewer.
}
