% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{view_run_details}
\alias{view_run_details}
\title{Initialize run details widget}
\usage{
view_run_details(run, auto_refresh = TRUE)
}
\arguments{
\item{run}{Run object}

\item{auto_refresh}{Boolean indicating whether or not widget should update
run details automatically. The default is TRUE when using RStudio.}
}
\description{
Initializes a ShinyApp in RStudio Viewer (or the default browser if Viewer
is unavailable) showing details of the submitted run. If using RStudio, the
plot will auto-update with information collected from the server. For more
details about the run, click the web view link. The widget will stop running
once the run has reached a terminal state: "Failed", "Completed", or
"Canceled".

If you are running this method from an RMarkdown file, the
run details table will show up in the code chunk output
instead of the Viewer.
}
