% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/webservice.R
\name{wait_for_deployment}
\alias{wait_for_deployment}
\title{Wait for a web service to finish deploying}
\usage{
wait_for_deployment(webservice, show_output = FALSE)
}
\arguments{
\item{webservice}{The \code{LocalWebservice}, \code{AciWebservice}, or
\code{AksWebservice} object.}

\item{show_output}{If \code{TRUE}, print more verbose output. Defaults
to \code{FALSE}.}
}
\value{
None
}
\description{
Automatically poll on the running web service deployment and
wait for the web service to reach a terminal state. Will throw
an exception if it reaches a non-successful terminal state.

Typically called after running \code{deploy_model()}.
}
\seealso{
\code{deploy_model()}
}
