% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute.R
\name{wait_for_provisioning_completion}
\alias{wait_for_provisioning_completion}
\title{Wait for a cluster to finish provisioning}
\usage{
wait_for_provisioning_completion(cluster, show_output = FALSE)
}
\arguments{
\item{cluster}{The \code{AmlCompute} or \code{AksCompute} object.}

\item{show_output}{If \code{TRUE}, more verbose output will be provided.}
}
\value{
None
}
\description{
Wait for a cluster to finish provisioning. Typically invoked after a
\code{create_aml_compute()} or \code{create_aks_compute()} call.
}
\seealso{
\code{create_aml_compute()}, \code{create_aks_compute()}
}
