% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{wait_for_run_completion}
\alias{wait_for_run_completion}
\title{Wait for the completion of a run}
\usage{
wait_for_run_completion(run, show_output = TRUE)
}
\arguments{
\item{run}{The \code{Run} object.}

\item{show_output}{If \code{TRUE}, print verbose output to console.}
}
\value{
None
}
\description{
Wait for the run to reach a terminal state. Typically called
after submitting an experiment run with \code{submit_experiment()}.
}
\seealso{
\code{\link[=submit_experiment]{submit_experiment()}}
}
